Move SE and BGM ids to a separate header
