+++ include/MSound/SoundEffects.hpp
@@ -1,0 +1,1635 @@
+// Exact order and names were pulled from mSound.asn
+enum MSoundSEId {
+	MSD_SE_PO_WATER_HI          = 0x0000,
+	MSD_SE_PO_WATER_LOW         = 0x0001,
+	MSD_SE_PO_ACTION_ON_EMPTY   = 0x0802,
+	MSD_SE_PO_PETBOTTLE_FLY     = 0x0003,
+	MSD_SE_PO_WT_DRY_UP         = 0x0804,
+	MSD_SE_PO_WATER_HI_TRG      = 0x0805,
+	MSD_SE_PO_WATER_LOW_TRG     = 0x0806,
+	MSD_SE_PO_HAND_OFF          = 0x0807,
+	MSD_SE_PO_NOZZLE_OFF        = 0x0808,
+	MSD_SE_PO_ACTION_ON         = 0x0809,
+	MSD_SE_PO_ACTION_OFF        = 0x080A,
+	MSD_SE_PO_FRICTION_A        = 0x080B,
+	MSD_SE_PO_FRICTION_B        = 0x080C,
+	MSD_SE_PO_FRICTION_C        = 0x080D,
+	MSD_SE_PO_SUCK_WATER_A      = 0x080E,
+	MSD_SE_PO_SUCK_WATER_B      = 0x000F,
+	MSD_SE_PO_ACTION_OFF_EMPTY  = 0x0810,
+	MSD_SE_SY_SELECT_POMP       = 0x0811,
+	MSD_SE_SY_SELECT_POMP_BACK  = 0x0812,
+	MSD_SE_PO_SHOTGUN_TRIGGER   = 0x0813,
+	MSD_SE_PO_SNIPER_TRIGGER    = 0x0814,
+	MSD_SE_PO_DOWN_OFF          = 0x0815,
+	MSD_SE_PO_POMPING1          = 0x0816,
+	MSD_SE_PO_POMPING2          = 0x0817,
+	MSD_SE_PO_HOVER             = 0x0018,
+	MSD_SE_PO_SPREAD            = 0x0019,
+	MSD_SE_PO_ROCKET_TRIGGER    = 0x081B,
+	MSD_SE_PO_LAUNCH_TRIGGER    = 0x081C,
+	MSD_SE_PO_WFLOWER_RISE_UP   = 0x001D,
+	MSD_SE_PO_WFLOWER_EXPLOSION = 0x081E,
+	MSD_SE_PO_SPRINKLER         = 0x001F,
+	MSD_SE_PO_GACHA1            = 0x0820,
+	MSD_SE_PO_GACHA2            = 0x0821,
+	MSD_SE_PO_GACHA3            = 0x0822,
+	MSD_SE_PO_GACHA4            = 0x0823,
+	MSD_SE_PO_NORMAL_NOZZLE_IMI = 0x0024,
+	MSD_SE_PO_GET_PETBOTTLE     = 0x0825,
+	MSD_SE_MA_SURF_START_ACCEL  = 0x0826,
+	MSD_SE_MA_SURF_TAME         = 0x0027,
+	MSD_SE_MA_SURF_JUMP         = 0x0828,
+	MSD_SE_MA_SURF_GND_ON_WATER = 0x0829,
+	MSD_SE_MA_TAKE_BARREL       = 0x1800,
+	MSD_SE_MA_SLIP_POLLUT       = 0x1001,
+	MSD_SE_MA_WALL_COL_SOFT     = 0x1802,
+	MSD_SE_MA_WALL_COL_HARD     = 0x1803,
+	MSD_SE_MA_FALL_IN_WATER_SLW = 0x1805,
+	MSD_SE_MA_FALL_IN_WATER_DEP = 0x1806,
+	MSD_SE_MA_KICK_EGG          = 0x1807,
+	MSD_SE_MA_POUR_WATER        = 0x1008,
+	MSD_SE_MA_SLIP              = 0x1009,
+	MSD_SE_MA_STALL_JUMP        = 0x180A,
+	MSD_SE_MA_SINK_IN_MUD       = 0x100B,
+	MSD_SE_MA_GET_ITEM          = 0x180C,
+	MSD_SE_MA_ROLL              = 0x180D,
+	MSD_SE_MA_HIP_ATTACK        = 0x180E,
+	MSD_SE_MA_ROPE_ROLL_FULL_B  = 0x180F,
+	MSD_SE_MA_ROPE_JUMP_A       = 0x1810,
+	MSD_SE_MA_ROPE_JUMP_B       = 0x1811,
+	MSD_SE_MA_ROPE_JUMP_C       = 0x1812,
+	MSD_SE_MA_DAMAGE_FIRE       = 0x1813,
+	MSD_SE_MA_DAMAGE_ELEC       = 0x1814,
+	MSD_SE_MA_ROPE_ROLL_HALF_A  = 0x1815,
+	MSD_SE_MA_ROPE_ROLL_HALF_B  = 0x1816,
+	MSD_SE_MA_ROPE_ROLL_FULL_A  = 0x1817,
+	MSD_SE_MA_KICK_ENEMY        = 0x1818,
+	MSD_SE_MA_KICK_ENEMY_2      = 0x1819,
+	MSD_SE_MA_KICK_ENEMY_3      = 0x181A,
+	MSD_SE_MA_KICK_ENEMY_4      = 0x181B,
+	MSD_SE_MA_KICK_ENEMY_5      = 0x181C,
+	MSD_SE_MA_JUMP_FR_WATER_SLW = 0x181D,
+	MSD_SE_MA_JUMP_FR_WATER_DEP = 0x181E,
+	MSD_SE_MA_DIVE_IN_MAHRE     = 0x101F,
+	MSD_SE_MA_WALK_STONE_L_HEEL = 0x1820,
+	MSD_SE_MA_WALK_STONE_L_TIP  = 0x1822,
+	MSD_SE_MA_WALK_STONE_R_HEEL = 0x1824,
+	MSD_SE_MA_WALK_STONE_R_TIP  = 0x1826,
+	MSD_SE_MA_WALK_STN_SND_LH   = 0x1828,
+	MSD_SE_MA_WALK_STN_SND_LT   = 0x182A,
+	MSD_SE_MA_WALK_STN_SND_RH   = 0x182C,
+	MSD_SE_MA_WALK_STN_SND_RT   = 0x182E,
+	MSD_SE_MA_WALK_MARBLE_L_H   = 0x1830,
+	MSD_SE_MA_WALK_MARBLE_L_T   = 0x1832,
+	MSD_SE_MA_WALK_MARBLE_R_H   = 0x1834,
+	MSD_SE_MA_WALK_MARBLE_R_T   = 0x1836,
+	MSD_SE_MA_WALK_SOIL_SLD_LH1 = 0x1838,
+	MSD_SE_MA_WALK_SOIL_SLD_LH2 = 0x1839,
+	MSD_SE_MA_WALK_SOIL_SLD_LT1 = 0x183A,
+	MSD_SE_MA_WALK_SOIL_SLD_LT2 = 0x183B,
+	MSD_SE_MA_WALK_SOIL_SLD_RH1 = 0x183C,
+	MSD_SE_MA_WALK_SOIL_SLD_RH2 = 0x183D,
+	MSD_SE_MA_WALK_SOIL_SLD_RT1 = 0x183E,
+	MSD_SE_MA_WALK_SOIL_SLD_RT2 = 0x183F,
+	MSD_SE_MA_WALK_SOIL_LH      = 0x1840,
+	MSD_SE_MA_WALK_SOIL_LT      = 0x1842,
+	MSD_SE_MA_WALK_SOIL_RH      = 0x1844,
+	MSD_SE_MA_WALK_SOIL_RT      = 0x1846,
+	MSD_SE_MA_WALK_SAND_L       = 0x1848,
+	MSD_SE_MA_WALK_SAND_R       = 0x184C,
+	MSD_SE_MA_WALK_GRAVEL_L     = 0x1850,
+	MSD_SE_MA_WALK_GRAVEL_R     = 0x1854,
+	MSD_SE_MA_WALK_WOODBOARD_LH = 0x1858,
+	MSD_SE_MA_WALK_WOODBOARD_LT = 0x185A,
+	MSD_SE_MA_WALK_WOODBOARD_RH = 0x185C,
+	MSD_SE_MA_WALK_WOODBOARD_RT = 0x185E,
+	MSD_SE_MA_WALK_WOOD_THN_LH1 = 0x1860,
+	MSD_SE_MA_WALK_WOOD_THN_LH2 = 0x1861,
+	MSD_SE_MA_WALK_WOOD_THN_LT1 = 0x1862,
+	MSD_SE_MA_WALK_WOOD_THN_LT2 = 0x1863,
+	MSD_SE_MA_WALK_WOOD_THN_RH1 = 0x1864,
+	MSD_SE_MA_WALK_WOOD_THN_RH2 = 0x1865,
+	MSD_SE_MA_WALK_WOOD_THN_RT1 = 0x1866,
+	MSD_SE_MA_WALK_WOOD_THN_RT2 = 0x1867,
+	MSD_SE_MA_WALK_WOOD_LH      = 0x1868,
+	MSD_SE_MA_WALK_WOOD_LT      = 0x186A,
+	MSD_SE_MA_WALK_WOOD_RH      = 0x186C,
+	MSD_SE_MA_WALK_WOOD_RT      = 0x186E,
+	MSD_SE_MA_WALK_WOOD_SQ_LH1  = 0x1870,
+	MSD_SE_MA_WALK_WOOD_SQ_LH2  = 0x1871,
+	MSD_SE_MA_WALK_WOOD_SQ_LT1  = 0x1872,
+	MSD_SE_MA_WALK_WOOD_SQ_LT2  = 0x1873,
+	MSD_SE_MA_WALK_WOOD_SQ_RH1  = 0x1874,
+	MSD_SE_MA_WALK_WOOD_SQ_RH2  = 0x1875,
+	MSD_SE_MA_WALK_WOOD_SQ_RT1  = 0x1876,
+	MSD_SE_MA_WALK_WOOD_SQ_RT2  = 0x1877,
+	MSD_SE_MA_WALK_METALNET_LH1 = 0x1878,
+	MSD_SE_MA_WALK_METALNET_LH2 = 0x1879,
+	MSD_SE_MA_WALK_METALNET_LT1 = 0x187A,
+	MSD_SE_MA_WALK_METALNET_LT2 = 0x187B,
+	MSD_SE_MA_WALK_METALNET_RH1 = 0x187C,
+	MSD_SE_MA_WALK_METALNET_RH2 = 0x187D,
+	MSD_SE_MA_WALK_METALNET_RT1 = 0x187E,
+	MSD_SE_MA_WALK_METALNET_RT2 = 0x187F,
+	MSD_SE_MA_WALK_METAL_VC_LH1 = 0x1880,
+	MSD_SE_MA_WALK_METAL_VC_LH2 = 0x1881,
+	MSD_SE_MA_WALK_METAL_VC_LT1 = 0x1882,
+	MSD_SE_MA_WALK_METAL_VC_LT2 = 0x1883,
+	MSD_SE_MA_WALK_METAL_VC_RH1 = 0x1884,
+	MSD_SE_MA_WALK_METAL_VC_RH2 = 0x1885,
+	MSD_SE_MA_WALK_METAL_VC_RT1 = 0x1886,
+	MSD_SE_MA_WALK_METAL_VC_RT2 = 0x1887,
+	MSD_SE_MA_WALK_METAL_SL_LH1 = 0x1888,
+	MSD_SE_MA_WALK_METAL_SL_LH2 = 0x1889,
+	MSD_SE_MA_WALK_METAL_SL_LT1 = 0x188A,
+	MSD_SE_MA_WALK_METAL_SL_LT2 = 0x188B,
+	MSD_SE_MA_WALK_METAL_SL_RH1 = 0x188C,
+	MSD_SE_MA_WALK_METAL_SL_RH2 = 0x188D,
+	MSD_SE_MA_WALK_METAL_SL_RT1 = 0x188E,
+	MSD_SE_MA_WALK_METAL_SL_RT2 = 0x188F,
+	MSD_SE_MA_WALK_BRANCH_L1    = 0x1890,
+	MSD_SE_MA_WALK_BRANCH_L2    = 0x1891,
+	MSD_SE_MA_WALK_BRANCH_R1    = 0x1894,
+	MSD_SE_MA_WALK_BRANCH_R2    = 0x1895,
+	MSD_SE_MA_WALK_TALLGRASS_L  = 0x1898,
+	MSD_SE_MA_WALK_TALLGRASS_L2 = 0x1899,
+	MSD_SE_MA_WALK_TALLGRASS_R  = 0x189C,
+	MSD_SE_MA_WALK_TALLGRASS_R2 = 0x189D,
+	MSD_SE_MA_WALK_LAWN_L       = 0x18A0,
+	MSD_SE_MA_WALK_LAWN_L2      = 0x18A1,
+	MSD_SE_MA_WALK_LAWN_R       = 0x18A4,
+	MSD_SE_MA_WALK_LAWN_R2      = 0x18A5,
+	MSD_SE_MA_WALK_STRAW_LH1    = 0x18A8,
+	MSD_SE_MA_WALK_STRAW_LH2    = 0x18A9,
+	MSD_SE_MA_WALK_STRAW_LT1    = 0x18AA,
+	MSD_SE_MA_WALK_STRAW_LT2    = 0x18AB,
+	MSD_SE_MA_WALK_STRAW_RH1    = 0x18AC,
+	MSD_SE_MA_WALK_STRAW_RH2    = 0x18AD,
+	MSD_SE_MA_WALK_STRAW_RT1    = 0x18AE,
+	MSD_SE_MA_WALK_STRAW_RT2    = 0x18AF,
+	MSD_SE_MA_WALK_ROOFTILE_LH  = 0x18B0,
+	MSD_SE_MA_WALK_ROOFTILE_LT  = 0x18B2,
+	MSD_SE_MA_WALK_ROOFTILE_RH  = 0x18B4,
+	MSD_SE_MA_WALK_ROOFTILE_RT  = 0x18B6,
+	MSD_SE_MA_WALK_ROOFTOTAN_LH = 0x18B8,
+	MSD_SE_MA_WALK_ROOFTOTAN_LT = 0x18BA,
+	MSD_SE_MA_WALK_ROOFTOTAN_RH = 0x18BC,
+	MSD_SE_MA_WALK_ROOFTOTAN_RT = 0x18BE,
+	MSD_SE_MA_WALK_ROOF_HOOD_L1 = 0x18C0,
+	MSD_SE_MA_WALK_ROOF_HOOD_L2 = 0x18C1,
+	MSD_SE_MA_WALK_ROOF_HOOD_R1 = 0x18C4,
+	MSD_SE_MA_WALK_ROOF_HOOD_R2 = 0x18C5,
+	MSD_SE_MA_WALK_WIRE_L1      = 0x18C8,
+	MSD_SE_MA_WALK_WIRE_L2      = 0x18C9,
+	MSD_SE_MA_WALK_WIRE_R1      = 0x18CC,
+	MSD_SE_MA_WALK_WIRE_R2      = 0x18CD,
+	MSD_SE_MA_WALK_TABLE_LH     = 0x18D0,
+	MSD_SE_MA_WALK_TABLE_LT     = 0x18D2,
+	MSD_SE_MA_WALK_TABLE_RH     = 0x18D4,
+	MSD_SE_MA_WALK_TABLE_RT     = 0x18D6,
+	MSD_SE_MA_WALK_BED_LH       = 0x18D8,
+	MSD_SE_MA_WALK_BED_LT       = 0x18DA,
+	MSD_SE_MA_WALK_BED_RH       = 0x18DC,
+	MSD_SE_MA_WALK_BED_RT       = 0x18DE,
+	MSD_SE_MA_WALK_CARPET_LH1   = 0x18E0,
+	MSD_SE_MA_WALK_CARPET_LH2   = 0x18E1,
+	MSD_SE_MA_WALK_CARPET_LT1   = 0x18E2,
+	MSD_SE_MA_WALK_CARPET_LT2   = 0x18E3,
+	MSD_SE_MA_WALK_CARPET_RH1   = 0x18E4,
+	MSD_SE_MA_WALK_CARPET_RH2   = 0x18E5,
+	MSD_SE_MA_WALK_CARPET_RT1   = 0x18E6,
+	MSD_SE_MA_WALK_CARPET_RT2   = 0x18E7,
+	MSD_SE_MA_WALK_CHAIR_LH1    = 0x18E8,
+	MSD_SE_MA_WALK_CHAIR_LH2    = 0x18E9,
+	MSD_SE_MA_WALK_CHAIR_LT1    = 0x18EA,
+	MSD_SE_MA_WALK_CHAIR_LT2    = 0x18EB,
+	MSD_SE_MA_WALK_CHAIR_RH1    = 0x18EC,
+	MSD_SE_MA_WALK_CHAIR_RH2    = 0x18ED,
+	MSD_SE_MA_WALK_CHAIR_RT1    = 0x18EE,
+	MSD_SE_MA_WALK_CHAIR_RT2    = 0x18EF,
+	MSD_SE_MA_WALK_GLASS_LH     = 0x18F8,
+	MSD_SE_MA_WALK_GLASS_LT     = 0x18FA,
+	MSD_SE_MA_WALK_GLASS_RH     = 0x18FC,
+	MSD_SE_MA_WALK_GLASS_RT     = 0x18FE,
+	MSD_SE_MA_WALK_KINOKO_L     = 0x1910,
+	MSD_SE_MA_WALK_KINOKO_R     = 0x1914,
+	MSD_SE_MA_WALK_CARPET2_LH1  = 0x1918,
+	MSD_SE_MA_WALK_CARPET2_LH2  = 0x1919,
+	MSD_SE_MA_WALK_CARPET2_LT1  = 0x191A,
+	MSD_SE_MA_WALK_CARPET2_LT2  = 0x191B,
+	MSD_SE_MA_WALK_CARPET2_RH1  = 0x191C,
+	MSD_SE_MA_WALK_CARPET2_RH2  = 0x191D,
+	MSD_SE_MA_WALK_CARPET2_RT1  = 0x191E,
+	MSD_SE_MA_WALK_CARPET2_RT2  = 0x191F,
+	MSD_SE_MA_SWIM              = 0x1920,
+	MSD_SE_MA_SWIM2             = 0x1921,
+	MSD_SE_MA_WALK_WATER_S_L    = 0x1924,
+	MSD_SE_MA_WALK_WATER_S_R    = 0x1926,
+	MSD_SE_MA_WALK_WATER_D_L    = 0x1928,
+	MSD_SE_MA_WALK_WATER_D_R    = 0x192A,
+	MSD_SE_MA_WALK_MUD_S_L      = 0x192C,
+	MSD_SE_MA_WALK_MUD_S_R      = 0x192E,
+	MSD_SE_MA_WALK_MUD_D_L      = 0x1930,
+	MSD_SE_MA_WALK_MUD_D_R      = 0x1932,
+	MSD_SE_MA_FALL_IN_WATER_VSL = 0x1938,
+	MSD_SE_MA_JUMP_FR_WATER_VSL = 0x1939,
+	MSD_SE_MA_FENCE_PUNCH       = 0x193A,
+	MSD_SE_MA_FENCE_CATCH       = 0x193B,
+	MSD_SE_MA_TIMP_HI           = 0x193C,
+	MSD_SE_MA_HANEAGARI         = 0x193D,
+	MSD_SE_MA_FALL_AFTER        = 0x193E,
+	MSD_SE_MA_SLIP_WALL         = 0x113F,
+	MSD_SE_MA_SLIP_TREE         = 0x1140,
+	MSD_SE_MA_SLIP_POLLUT_CP    = 0x1141,
+	MSD_SE_YO_WALK_L            = 0x1942,
+	MSD_SE_YO_WALK_R            = 0x1943,
+	MSD_SE_YO_BORN              = 0x1944,
+	MSD_SE_YO_TONGUE_BACK       = 0x1945,
+	MSD_SE_YO_TONGUE_RELEASE    = 0x1946,
+	MSD_SE_YO_TONGUE_GOKKUN     = 0x1947,
+	MSD_SE_MA_WALL_COL_CMN_H    = 0x1948,
+	MSD_SE_MA_WALL_COL_CMN_S    = 0x1949,
+	MSD_SE_MA_WALL_COL_CMN_VS   = 0x194A,
+	MSD_SE_MA_WALL_COL_VERYSOFT = 0x194B,
+	MSD_SE_MA_PUT_BARREL        = 0x194C,
+	MSD_SE_MA_TAKE_DRIAN        = 0x194D,
+	MSD_SE_MA_PUT_DRIAN         = 0x194E,
+	MSD_SE_MA_KICK_DRIAN        = 0x194F,
+	MSD_SE_MA_WATER_WAIT        = 0x1950,
+	MSD_SE_MA_WATER_WAIT2       = 0x1951,
+	MSD_SE_MA_SWIM_MOVE3        = 0x1952,
+	MSD_SE_MA_SWIM_FOOT2        = 0x1953,
+	MSD_SE_MA_FLOAT             = 0x1954,
+	MSD_SE_MA_FLOAT2            = 0x1955,
+	MSD_SE_MA_SWIM_MOVE         = 0x1956,
+	MSD_SE_MA_SWIM_MOVE2        = 0x1957,
+	MSD_SE_MA_TIMP_LOW          = 0x1958,
+	MSD_SE_MA_TIMP_MID          = 0x1959,
+	MSD_SE_MA_BOUND             = 0x195A,
+	MSD_SE_MA_Y_ROLL            = 0x195B,
+	MSD_SE_MA_Y_ROLL2           = 0x195C,
+	MSD_SE_MA_Y_ROLL3           = 0x195D,
+	MSD_SE_MA_Y_ROLL4           = 0x195E,
+	MSD_SE_MA_Y_ROLL5           = 0x195F,
+	MSD_SE_MA_Y_ROLL6           = 0x1960,
+	MSD_SE_MA_Y_ROLL7           = 0x1961,
+	MSD_SE_MA_Y_ROLL8           = 0x1962,
+	MSD_SE_MA_Y_ROLL9           = 0x1963,
+	MSD_SE_MA_Y_ROLL10          = 0x1964,
+	MSD_SE_MA_WALL_COL_HEAD     = 0x1965,
+	MSD_SE_MA_INTO_BOTTLE       = 0x1966,
+	MSD_SE_MA_BOTTLE_WIND       = 0x1967,
+	MSD_SE_MA_HELMET            = 0x1968,
+	MSD_SE_MA_ROLL_C            = 0x1969,
+	MSD_SE_MA_ROLL_D            = 0x196A,
+	MSD_SE_MA_ROLL_E            = 0x196B,
+	MSD_SE_MA_ROLL_F            = 0x196C,
+	MSD_SE_MA_ROLL_G            = 0x196D,
+	MSD_SE_MA_ROLL_A            = 0x196E,
+	MSD_SE_MA_ROLL_B            = 0x196F,
+	MSD_SE_MA_ROLL_CU           = 0x1970,
+	MSD_SE_MA_BELT              = 0x1971,
+	MSD_SE_MA_SIRIMOTHI         = 0x1972,
+	MSD_SE_MA_ROLL_GD           = 0x1973,
+	MSD_SE_MA_ROLL_AD           = 0x1974,
+	MSD_SE_MA_ROLL_BD           = 0x1975,
+	MSD_SE_MA_KAGE_FIELD_AWAY   = 0x1976,
+	MSD_SE_MA_KAGE_GATE_AWAY    = 0x1977,
+	MSD_SE_MA_KAGE_TAKE_PEACH   = 0x1978,
+	MSD_SE_MA_WARP              = 0x1979,
+	MSD_SE_MA_WARP_EX           = 0x197A,
+	MSD_SE_MA_UMARI             = 0x197B,
+	MSD_SE_MA_KAGE_FIELD_APPEAR = 0x197C,
+	MSD_SE_MA_SURF_WATER        = 0x117D,
+	MSD_SE_MA_SURF_AIR          = 0x117E,
+	MSD_SE_MA_SURF_SLIP         = 0x117F,
+	MSD_SE_MA_KAGE_PAINTS       = 0x1980,
+	MSD_SE_MA_KAGE_PAINTINGWIND = 0x1981,
+	MSD_SE_MA_HAT_STOLEN        = 0x1982,
+	MSD_SE_MA_TAKE_HAT          = 0x1983,
+	MSD_SE_MA_SHINING_EFF1      = 0x1984,
+	MSD_SE_MA_SHINING_EFF2      = 0x1985,
+	MSD_SE_MA_SHINING_EFF3      = 0x1986,
+	MSD_SE_MA_SHINING_EFF4      = 0x1987,
+	MSD_SE_MA_SHINING_EFF_CALM1 = 0x1988,
+	MSD_SE_MA_SHINING_EFF_CALM2 = 0x1989,
+	MSD_SE_MA_SHINING_EFF_CALM3 = 0x198A,
+	MSD_SE_MA_SHINING_EFF_CALM4 = 0x198B,
+	MSD_SE_YO_AWAY              = 0x198C,
+	MSD_SE_YO_HIP_ATTACK        = 0x198D,
+	MSD_SE_MA_ROLL_JUMP         = 0x198E,
+	MSD_SE_EN_TYOROBE_BYE       = 0x1990,
+	MSD_SE2_OBJ_CORK_WEEP       = 0x1991,
+	MSD_SE2_EN_TYUHANA_JUMP     = 0x1992,
+	MSD_SE2_EN_TAMANOKO_SHOOK   = 0x1993,
+	MSD_SE2_OBJ_FALL_JET        = 0x1994,
+	MSD_SE2_EN_TYOROBE_V_DOWN   = 0x1995,
+	MSD_SE2_EN_BOMBHEI_BOMB     = 0x1996,
+	MSD_SE_OBJ_FOUNTAIN         = 0x3000,
+	MSD_SE_IT_EGG_APPEAR        = 0x3801,
+	MSD_SE_IT_COMMON_APPEAR     = 0x3802,
+	MSD_SE_EF_FIRE              = 0x3803,
+	MSD_SE_IT_EGG_BOUND         = 0x3804,
+	MSD_SE_EF_ELEC              = 0x3805,
+	MSD_SE_EF_LIGHTNING         = 0x3806,
+	MSD_SE_GE_FALL              = 0x3007,
+	MSD_SE_OBJ_QUAKE            = 0x3008,
+	MSD_SE_IT_SOCCER_GOAL       = 0x3809,
+	MSD_SE_IT_BARREL_CRASH      = 0x380A,
+	MSD_SE_IT_BARREL_ROLL1      = 0x380B,
+	MSD_SE_IT_BARREL_ROLL2      = 0x380C,
+	MSD_SE_OBJ_STAND_BREAK      = 0x300D,
+	MSD_SE_OBJ_BATH_BREAK       = 0x300E,
+	MSD_SE_OBJ_TREE_APPEAR      = 0x300F,
+	MSD_SE_OBJ_POLE_APPEAR      = 0x3010,
+	MSD_SE_NPC_APPEAR           = 0x3811,
+	MSD_SE_IT_APPEAR            = 0x3812,
+	MSD_SE_OBJ_KAMOME_SOLO      = 0x3813,
+	MSD_SE_OBJ_KAMOME_SOLO_12   = 0x3814,
+	MSD_SE_OBJ_KAMOME_SOLO_13   = 0x3815,
+	MSD_SE_OBJ_KAMOME_SOLO_21   = 0x3816,
+	MSD_SE_OBJ_KAMOME_SOLO_22   = 0x3817,
+	MSD_SE_OBJ_KAMOME_SOLO_23   = 0x3818,
+	MSD_SE_OBJ_KAMOME_ENSB      = 0x3819,
+	MSD_SE_OBJ_KAMOME_ENSB_2    = 0x381A,
+	MSD_SE_OBJ_KAMOME_ENSB_3    = 0x381B,
+	MSD_SE_OBJ_ROPE_CLEAK_A     = 0x381C,
+	MSD_SE_OBJ_ROPE_CLEAK_B     = 0x381D,
+	MSD_SE_OBJ_ROPE_CLEAK_ROLL  = 0x381E,
+	MSD_SE_OBJ_ROPE_CLEAK_HALFA = 0x381F,
+	MSD_SE_OBJ_ROPE_CLEAK_HALFB = 0x3820,
+	MSD_SE_OBJ_SUPERBLOCK_BREAK = 0x3821,
+	MSD_SE_OBJ_FALL_SMALL       = 0x3022,
+	MSD_SE_OBJ_SUBMARINE_SINK   = 0x3023,
+	MSD_SE_OBJ_FENCE_REVERSE1   = 0x3824,
+	MSD_SE_OBJ_FENCE_REVERSE2   = 0x3825,
+	MSD_XX_OBJ_FLOWER_OPEN_0    = 0x3826,
+	MSD_XX_OBJ_FLOWER_OPEN_1    = 0x3827,
+	MSD_XX_OBJ_FLOWER_OPEN_2    = 0x3828,
+	MSD_XX_OBJ_FLOWER_OPEN_3    = 0x3829,
+	MSD_XX_OBJ_FLOWER_OPEN_4    = 0x382A,
+	MSD_XX_OBJ_FLOWER_OPEN_5    = 0x382B,
+	MSD_XX_OBJ_FLOWER_OPEN_6    = 0x382C,
+	MSD_XX_OBJ_FLOWER_OPEN_7    = 0x382D,
+	MSD_SE_OBJ_PONPONSEN        = 0x302E,
+	MSD_SE_OBJ_FLAG             = 0x302F,
+	MSD_SE_OBJ_SUBMARINE_CRANE  = 0x3030,
+	MSD_SE_OBJ_SUBMARINE_MILL   = 0x3031,
+	MSD_SE_OBJ_SUBMARINE_SET    = 0x3832,
+	SMD_SE_OBJ_SUBMARINE_SET_WT = 0x3833,
+	MSD_SE_OBJ_CRANE_SIDEMOVE1  = 0x3034,
+	MSD_SE_OBJ_CRANE_SIDEMOVE2  = 0x3035,
+	MSD_SE_OBJ_CRANE_UPDOWN1    = 0x3036,
+	MSD_SE_OBJ_CRANE_UPDOWN2    = 0x3037,
+	MSD_SE_OBJ_WATER_FENCE_FW   = 0x3838,
+	MSD_SE_OBJ_WATER_FENCE_REV  = 0x3839,
+	MSD_SE_OBJ_MANHOLE_OPEN     = 0x383B,
+	MSD_SE_OBJ_MANHOLE_CLOSE    = 0x383C,
+	MSD_SE_OBJ_MANHOLE_DOWN     = 0x383D,
+	MSD_SE_OBJ_MANHOLE_UP       = 0x383E,
+	MSD_SE_OBJ_BI_STEPMILL_WIND = 0x3040,
+	MSD_XX_OBJ_BI_BELL          = 0x3841,
+	MSD_SE_OBJ_BI_STEPMILL_MOVE = 0x3042,
+	MSD_SE_OBJ_BI_BIGMILL       = 0x3043,
+	MSD_SE_OBJ_BI_UPDOWNMILL    = 0x3044,
+	MSD_SE_OBJ_BI_COINMILL      = 0x3045,
+	MSD_SE_OBJ_LEV_DOWN_TEST    = 0x3046,
+	MSD_SE_OBJ_BI_WINDMILLTOWER = 0x3047,
+	MSD_SE_OBJ_MA_MIRROR_MOVE   = 0x3048,
+	MSD_SE_OBJ_MA_MIRROR_IMPACT = 0x3849,
+	MSD_SE_OBJ_MA_MIRROR_DEMOMV = 0x304A,
+	MSD_SE_OBJ_SAND_GROW        = 0x304B,
+	MSD_SE_OBJ_AP_BUTTON        = 0x384C,
+	MSD_SE_OBJ_AP_WATER_STREAM  = 0x304D,
+	MSD_SE_OBJ_BASKET_BOUND     = 0x384E,
+	MSD_SE_OBJ_BILLBOARD_MOVE   = 0x384F,
+	MSD_SE_OBJ_KUGI_IMPACT      = 0x3850,
+	MSD_SE_CHEST_ROLL           = 0x3851,
+	MSD_SE_CHEST_ROLL_REVERSE   = 0x3852,
+	MSD_SE_OBJ_DOOR_OPEN        = 0x3853,
+	MSD_SE_OBJ_DOOR_OPEN_KICK   = 0x3854,
+	MSD_SE_OBJ_DOOR_CLOSE       = 0x3855,
+	MSD_SE_OBJ_DOOR_CLOSE_KICK  = 0x3856,
+	MSD_SE_OBJ_GLASS_BREAK      = 0x3857,
+	MSD_SE_OBJ_PINNADOOR_OPEN_L = 0x3058,
+	MSD_SE_OBJ_PINNADOOR_IMPACT = 0x3859,
+	MSD_SE_OBJ_JET_COASTER      = 0x305A,
+	MSD_SE_OBJ_JET_COASTER_IMI  = 0x305B,
+	MSD_SE_OBJ_PANEL_BREAK      = 0x385C,
+	MSD_SE_OBJ_PANEL_ROLL       = 0x385D,
+	MSD_SE_OBJ_PUNCHER_APPEAR   = 0x385E,
+	MSD_SE_OBJ_PUNCHER_RETURN   = 0x385F,
+	MSD_SE_OBJ_MR_TSUBO_PULL    = 0x3060,
+	MSD_SE_OBJ_MR_TSUBO_WATER   = 0x3061,
+	MSD_SE_IT_DRIAN_BOUND       = 0x3862,
+	MSD_SE_OBJ_MVING_FENCT_SET  = 0x3863,
+	MSD_SE_OBJ_MVING_FENCT_PNCH = 0x3864,
+	MSD_SE_OBJ_MVING_FENCE_MOVE = 0x3065,
+	MSD_SE_OBJ_JUMP_KINOKO      = 0x3866,
+	MSD_SE_OBJ_SWING1           = 0x3867,
+	MSD_SE_OBJ_SWING2           = 0x3868,
+	MSD_SE_OBJ_BIRD_DOL_FLYING1 = 0x3869,
+	MSD_SE_OBJ_BIRD_DOL_FLYING2 = 0x386A,
+	MSD_SE_OBJ_BIRD_DOL_TO_FLY1 = 0x386B,
+	MSD_SE_OBJ_BIRD_DOL_TO_FLY2 = 0x386C,
+	MSD_SE_OBJ_BIRD_MAH_A1      = 0x386D,
+	MSD_SE_OBJ_BIRD_MAH_A2      = 0x386E,
+	MSD_SE_OBJ_BIRD_MAH_A3      = 0x386F,
+	MSD_SE_OBJ_BIRD_DOL_CHUN    = 0x3870,
+	MSD_SE_OBJ_BIRD_WING1       = 0x3871,
+	MSD_SE_OBJ_BIRD_WING2       = 0x3872,
+	MSD_SE_OBJ_BIRD_WING3       = 0x3873,
+	MSD_SE_OBJ_BIRD_WING4       = 0x3874,
+	MSD_SE_OBJ_DRINA_TO_WATER   = 0x3875,
+	MSD_SE_OBJ_GATE_LIGHT       = 0x3076,
+	MSD_SE_OBJ_GATE_WAVE        = 0x3077,
+	MSD_SE_OBJ_CLASSIC_BLOCK_B  = 0x3878,
+	MSD_SE_OBJ_ICE_BLOCK_MELT   = 0x3079,
+	MSD_XX_SHINE_EXIST          = 0x307A,
+	MSD_SE_YV_SHINE_GET         = 0x387B,
+	MSD_SE_OBJ_DORO_BROKEN      = 0x387C,
+	MSD_SE_SMOKE_EFFECT         = 0x387D,
+	MSD_SE_OBJ_DORO_GURAGURA    = 0x387E,
+	MSD_SE_OBJ_DORO_GURA_WATER  = 0x387F,
+	MSD_SE_OBJ_DORO_FLOAT       = 0x3080,
+	MSD_SE_OBJ_AWAY_INTO_GRAF   = 0x3881,
+	MSD_SE_OBJ_DOOR_LOCK        = 0x3882,
+	MSD_SE_OBJ_DOOR_GACHA       = 0x3883,
+	MSD_SE_OBJ_WATAGE_WIND      = 0x3884,
+	MSD_SE_OBJ_MAHRE_GATE_LIGHT = 0x3085,
+	MSD_SE_OBJ_KANRANSYA_RAPID  = 0x3086,
+	MSD_SE_OBJ_WATERMELON_MACN  = 0x3887,
+	MSD_SE_OBJ_WATERMELON_CRASH = 0x3888,
+	MSD_SE_OBJ_WATERMELON_BBUND = 0x3889,
+	MSD_SE_OBJ_WATERMELON_BROLL = 0x308A,
+	MSD_SE_OBJ_WATERMELON_SROLL = 0x308B,
+	MSD_SE_OBJ_WATERMELON_SBUND = 0x388C,
+	MSD_SE_OBJ_SLOT_SPIN        = 0x308D,
+	MSD_SE_OBJ_SLOT_INC_L       = 0x388E,
+	MSD_SE_OBJ_SLOT_INC_R       = 0x388F,
+	MSD_SE_OBJ_SLOT_INC_C       = 0x3890,
+	MSD_SE_OBJ_WIND_C           = 0x3891,
+	MSD_SE_OBJ_WIND_D           = 0x3892,
+	MSD_SE_OBJ_WIND_E           = 0x3893,
+	MSD_SE_OBJ_WIND_F           = 0x3894,
+	MSD_SE_OBJ_WIND_G           = 0x3895,
+	MSD_SE_OBJ_WIND_A           = 0x3896,
+	MSD_SE_OBJ_WIND_B           = 0x3897,
+	MSD_SE_OBJ_WIND_CU          = 0x3898,
+	MSD_SE_OBJ_WIND_GL          = 0x3899,
+	MSD_SE_OBJ_WIND_AL          = 0x389A,
+	MSD_SE_OBJ_WIND_BL          = 0x389B,
+	MSD_SE_OBJ_TEL_CLOSET_ROLL  = 0x389C,
+	MSD_SE_OBJ_TEL_CLOSET_RAPID = 0x389D,
+	MSD_SE_OBJ_PANELPUZZLE_WIND = 0x389E,
+	MSD_SE_OBJ_PIN_SHELL_CLOSE  = 0x389F,
+	MSD_SE_OBJ_PIN_BIKING_WING  = 0x38A0,
+	MSD_SE_OBJ_PIN_SHELL_OPEN   = 0x38A1,
+	MSD_SE_GE_FALL_UPPER        = 0x30A2,
+	MSD_SE_OBJ_WATERMELON_BLOCK = 0x38A3,
+	MSD_SE_OBJ_DOL_BEL_GS4_E4   = 0x38A4,
+	MSD_SE_OBJ_DOL_BEL_GS4      = 0x38A5,
+	MSD_SE_OBJ_DOL_BEL_E4       = 0x38A6,
+	MSD_SE_OBJ_DOL_BEL_GS4_E4_K = 0x38A7,
+	MSD_SE_OBJ_DOL_BEL_GS4_K    = 0x38A8,
+	MSD_SE_OBJ_DOL_BEL_E4_K     = 0x38A9,
+	MSD_SE_OBJ_SANDBLOCK_BREAK  = 0x30AA,
+	MSD_SE_OBJ_BBFENCE_PUNCH    = 0x38AB,
+	MSD_SE_OBJ_BBFENCE_CATCH    = 0x38AC,
+	MSD_SE_OBJ_BBFENCE_TURN1    = 0x38AD,
+	MSD_SE_OBJ_BBFENCE_RUN_STOP = 0x38AE,
+	MSD_SE_OBJ_BBFENCE_TURN2    = 0x38AF,
+	MSD_SE_DM_OSEN_DISAPPEAR    = 0x38B0,
+	MSD_SE_ENV_FALL_JET_LEVEL   = 0x30B1,
+	MSD_SE_OBJ_CALM_FLAME       = 0x30B2,
+	MSD_SE_EN_CANNON_DOWN       = 0x38B3,
+	MSD_SE_EN_CANNON_LAND       = 0x38B4,
+	MSD_SE_BS_TELESA_RLT_MOVE3  = 0x30B5,
+	MSD_SE_BS_TELESA_RLT_MOVE2  = 0x30B6,
+	MSD_SE_BS_TELESA_RLT_MOVE1  = 0x30B7,
+	MSD_SE_OBJ_BIA_BELL_SINGLE  = 0x38B8,
+	MSD_SE_SY_HP_EMPTY          = 0x4800,
+	MSD_SE_SY_HP_RECOVER        = 0x4801,
+	MSD_SE_SY_PAUSE_ON          = 0x4802,
+	MSD_SE_SY_PAUSE_OFF         = 0x4803,
+	MSD_SE_SY_TALK_MODE_IN      = 0x4804,
+	MSD_SE_SY_TALK_MODE_OUT     = 0x4805,
+	MSD_SE_SY_PO_AIR_PRESSURE   = 0x4006,
+	MSD_SE_SY_PO_WATER_FULL     = 0x4807,
+	MSD_SE_SY_PO_AIR_P_FULLDOWN = 0x4808,
+	MSD_SE_SY_PO_AIR_P_DOWN     = 0x4809,
+	MSD_SE_FGM_SOCCER_GOAL      = 0x480A,
+	MSD_SE_FGM_RESET_OBJECT     = 0x480B,
+	MSD_SE_SY_HP_DOWN           = 0x480C,
+	MSD_SE_SY_OBJ_APPEAR        = 0x480D,
+	MSD_SE_SY_CURSOR            = 0x480F,
+	MSD_SE_SY_DECIDE            = 0x4810,
+	MSD_SE_SY_COIN              = 0x4811,
+	MSD_SE_SY_NEWS              = 0x4812,
+	MSD_SE_SY_COIN_APPEAR       = 0x4813,
+	MSD_SE_SY_GET_SHINE         = 0x4814,
+	MSD_SE_SY_CURSOR_COMMON     = 0x4815,
+	MSD_SE_SY_WIPE_IN           = 0x4817,
+	MSD_SE_SY_WIPE_OUT          = 0x4818,
+	MSD_SE_SY_2D_IN             = 0x4819,
+	MSD_SE_SY_2D_OUT            = 0x481A,
+	MSD_SE_SY_DECIDE_COMMON     = 0x481B,
+	MSD_SE_SY_SELECT_COMMON     = 0x481C,
+	MSD_SE_SY_CANCEL_COMMON     = 0x481D,
+	MSD_SE_SY_E3_MENU_CURSOR    = 0x481E,
+	MSD_SE_CLEAN_UP_GRAF_OLD    = 0x481F,
+	MSD_SE_SY_MENU_SHINE_LIGHT  = 0x4820,
+	MSD_SE_SHINE_APPEAR         = 0x4821,
+	MSD_SE_SY_NEWP_AIR_TAME     = 0x4022,
+	MSD_SE_SY_DAMAGE            = 0x4823,
+	MSD_SE_SY_CAMERA_UP         = 0x4824,
+	MSD_SE_SY_CAMERA_OUT        = 0x4825,
+	MSD_SE_SY_CAMERA_BACK       = 0x4826,
+	MSD_SE_SY_TALK_OBI          = 0x4827,
+	MSD_SE_SY_CHANGE_TO_DOLBY   = 0x4828,
+	MSD_SE_SY_CHANGE_TO_STEREO  = 0x4829,
+	MSD_SE_SY_CHANGE_TO_MONO    = 0x482A,
+	MSD_SE_SY_HP_IN             = 0x482B,
+	MSD_SE_SY_HP_OUT            = 0x482C,
+	MSD_SE_SY_HP_RECOVER_MAX    = 0x482D,
+	MSD_SE_SY_E3_MENU_CURSOR1   = 0x482E,
+	MSD_SE_SY_E3_MENU_CURSOR2   = 0x482F,
+	MSD_SE_SY_E3_MENU_CURSOR3   = 0x4830,
+	MSD_SE_SY_E3_MENU_CURSOR4   = 0x4831,
+	MSD_SE_SY_E3_MENU_CURSOR5   = 0x4832,
+	MSD_SE_SY_DOLBY_TEST        = 0x4834,
+	MSD_SE_SY_TIMER_SIMPLE      = 0x4035,
+	MSD_SE_SY_TIMER_HURRY_UP    = 0x4036,
+	MSD_SE_SY_TIMER_HURRY_UP2   = 0x4037,
+	MSD_SE_SY_TIMER_HURRY_UP3   = 0x4038,
+	MSD_SE_SY_TIMER_HURRY_UP4   = 0x4039,
+	MSD_SE_SY_TIMER             = 0x403A,
+	MSD_SE_SY_DAMAGE_MID        = 0x483B,
+	MSD_SE_SY_DAMAGE_LIGHT      = 0x483C,
+	MSD_SE_SY_NOT_COLLECT       = 0x483D,
+	MSD_SE_SY_NOT_COLLECT_YOSHI = 0x483E,
+	MSD_SE_SY_COLLECT_PRETTY    = 0x483F,
+	MSD_SE_SY_COLLECT_YOSHI     = 0x4840,
+	MSD_SE_SY_1UP               = 0x4841,
+	MSD_SE_SY_COIN_BOUND        = 0x4842,
+	MSD_SE_SY_TIMECOIN_APPEAR   = 0x4843,
+	MSD_SE_SY_SUCCESS_MELODIAS  = 0x4844,
+	MSD_SE_SY_BLUE_COIN_GET     = 0x4845,
+	MSD_SE_SY_RED_COIN_GET      = 0x4846,
+	MSD_SE_SY_PROBLEM_SOLVED    = 0x4847,
+	MSD_SE_SY_PROBLEM_SIGN      = 0x4848,
+	MSD_SE_SY_COLLECT_DELIGHT   = 0x4849,
+	MSD_SE_SY_CLEAR_SIGN        = 0x484A,
+	MSD_SE_SY_TIMEEVENT_FAILED  = 0x484B,
+	MSD_SE_SY_COIN_DISAPPEAR    = 0x484C,
+	MSD_SE_SY_CLEAR_SIGN_BIG    = 0x484D,
+	MSD_SE_SY_GET_NOZZLE        = 0x484E,
+	MSD_SE_SY_GET_POMP          = 0x484F,
+	MSD_SE_SY_GET_COMMON_ITEM   = 0x4850,
+	MSD_SE_SY_RACE_START        = 0x4851,
+	MSD_SE_SY_RACE_FIRE         = 0x4852,
+	MSD_SE_SY_DAMAGE_WITH_LOSTH = 0x4853,
+	MSD_SE_SY_1UP_APPEAR        = 0x4854,
+	MSD_SE_SY_SHINE_SELECT      = 0x4855,
+	MSD_SE_SY_SHINE_CURSOR      = 0x4856,
+	MSD_SE_SY_DONCHO_OPEN       = 0x4857,
+	MSD_SE_SY_PANELPUZZLE_OPEN  = 0x4058,
+	MSD_SE_MA_INTO_DOKAN        = 0x4859,
+	MSD_SE_EV_GLOBAL_SEA_L      = 0x5000,
+	MSD_SE_EV_GLOBAL_SEA_R      = 0x5001,
+	MSD_SE_EV_GLOBAL_SEA_SR     = 0x5002,
+	MSD_SE_EV_LITTLE_WAVE_L     = 0x5003,
+	MSD_SE_EV_LITTLE_WAVE_R     = 0x5004,
+	MSD_SE_EV_LITTLE_WAVE_SR    = 0x5005,
+	MSD_SE_EV_ARCHED_FOUNTAIN   = 0x5006,
+	MSD_SE_EV_VTINY_WAVE_2_L    = 0x5009,
+	MSD_SE_EV_VTINY_WAVE_2_R    = 0x500A,
+	MSD_SE_EV_VTINY_WAVE_2_SR   = 0x500B,
+	MSD_SE_EV_TINY_WAVE_3_L     = 0x500C,
+	MSD_SE_EV_TINY_WAVE_3_R     = 0x500D,
+	MSD_SE_EV_TINY_WAVE_3_SR    = 0x500E,
+	MSD_SE_EV_STREAM            = 0x500F,
+	MSD_SE_EV_STREAM_DOWN       = 0x5010,
+	MSD_SE_EV_MONTE_WIND        = 0x5011,
+	MSD_SE_EV_CORONA            = 0x5012,
+	MSD_SE_OBJ_BIRD_BIA_1       = 0x5814,
+	MSD_SE_OBJ_BIRD_BIA_2       = 0x5815,
+	MSD_SE_OBJ_BIRD_BIA_3       = 0x5816,
+	MSD_SE_OBJ_BIRD_BIA_4       = 0x5817,
+	MSD_SE_OBJ_BIRD_BIA_5       = 0x5818,
+	MSD_SE_OBJ_BIRD_BIA_6       = 0x5819,
+	MSD_SE_OBJ_MONTE_DAY_A1     = 0x581B,
+	MSD_SE_OBJ_MONTE_DAY_A2     = 0x581C,
+	MSD_SE_OBJ_MONTE_DAY_A3     = 0x581D,
+	MSD_SE_OBJ_MONTE_DAY_B      = 0x581E,
+	MSD_SE_OBJ_MONTE_DAY_C      = 0x581F,
+	MSD_SE_OBJ_MONTE_NIGHT_A1   = 0x5820,
+	MSD_SE_OBJ_MONTE_NIGHT_A2   = 0x5821,
+	MSD_SE_OBJ_MONTE_NIGHT_A3   = 0x5822,
+	MSD_SE_ENV_MONTE_UPPER_WIND = 0x5023,
+	MSD_SE_WT_GND               = 0x6800,
+	MSD_SE_WT_INTO_WATER        = 0x6801,
+	MSD_SE_EN_COMMON_W_HIT_OK   = 0x6802,
+	MSD_SE_WT_DRY_UP            = 0x6806,
+	MSD_SE_HINOKURI_LIQUID_GND  = 0x6807,
+	MSD_SE_HINOKURI_LIQUID_GND2 = 0x6808,
+	MSD_SE_ERASE_SCRAWL         = 0x6809,
+	MSD_SE_ERASE_SCRAWL_CONT    = 0x680A,
+	MSD_SE_WT_BOSS_FADEAWAY     = 0x6010,
+	MSD_SE_MA_VO_JUMP_BIG_0     = 0x7800,
+	MSD_SE_MA_VO_JUMP_BIG_1     = 0x7801,
+	MSD_SE_MA_VO_JUMP_BIG_2     = 0x7802,
+	MSD_SE_MA_VO_JUMP_MID_0     = 0x7803,
+	MSD_SE_MA_VO_JUMP_MID_1     = 0x7804,
+	MSD_SE_MA_VO_JUMP_MID_2     = 0x7805,
+	MSD_SE_MA_VO_JUMP_MID_3     = 0x7806,
+	MSD_SE_MA_VO_JUMP_SMALL_0   = 0x7807,
+	MSD_SE_MA_VO_JUMP_SMALL_1   = 0x7808,
+	MSD_SE_MA_VO_JUMP_SMALL_2   = 0x7809,
+	MSD_SE_MA_VO_JUMP_SPECIAL_0 = 0x780A,
+	MSD_SE_MA_VO_JUMP_SPECIAL_1 = 0x780B,
+	MSD_SE_MA_VO_JUMP_SPECIAL_2 = 0x780C,
+	MSD_SE_MA_VO_JUMP_SPECIAL_3 = 0x780D,
+	MSD_SE_MA_VO_LAND_HI_0      = 0x780E,
+	MSD_SE_MA_VO_LAND_HI_1      = 0x780F,
+	MSD_SE_MA_VO_LAND_HI_2      = 0x7810,
+	MSD_SE_MA_VO_LAND_HI_3      = 0x7811,
+	MSD_SE_MA_VO_LAND_HI_4      = 0x7812,
+	MSD_SE_MA_VO_LAND_LOW_0     = 0x7813,
+	MSD_SE_MA_VO_LAND_LOW_1     = 0x7814,
+	MSD_SE_MA_VO_LAND_LOW_2     = 0x7815,
+	MSD_SE_MA_VO_LAND_LOW_3     = 0x7816,
+	MSD_SE_MA_VO_LAND_LOW_4     = 0x7817,
+	MSD_SE_MA_VO_DAMAGE_BIG     = 0x7818,
+	MSD_SE_MA_VO_DAMAGE_BIG_1   = 0x7819,
+	MSD_SE_MA_VO_DAMAGE_BIG_2   = 0x781A,
+	MSD_SE_MA_VO_DEAD           = 0x781B,
+	MSD_SE_MA_VO_DAMAGE_SMALL   = 0x781C,
+	MSD_SE_MA_VO_DAMAGE_SMALL_1 = 0x781D,
+	MSD_SE_MA_VO_DAMAGE_SMALL_2 = 0x781E,
+	MSD_SE_MA_VO_FOO            = 0x781F,
+	MSD_SE_MA_VO_FOO_1          = 0x7820,
+	MSD_SE_MV01_DAMAGE_BIG_01   = 0x7830,
+	MSD_SE_MV01_DAMAGE_BIG_02   = 0x7831,
+	MSD_SE_MV01_DAMAGE_BIG_03   = 0x7832,
+	MSD_SE_MV02_DAMAGE_MID_01   = 0x7833,
+	MSD_SE_MV02_DAMAGE_MID_02   = 0x7834,
+	MSD_SE_MV02_DAMAGE_MID_03   = 0x7835,
+	MSD_SE_MV02_DAMAGE_MID_04   = 0x7836,
+	MSD_SE_MV03_DAMAGE_LITLE_01 = 0x783B,
+	MSD_SE_MV03_DAMAGE_LITLE_02 = 0x783C,
+	MSD_SE_MV03_DAMAGE_LITLE_03 = 0x783D,
+	MSD_SE_MV03_DAMAGE_LITLE_04 = 0x783E,
+	MSD_SE_MV04_DAMAGE_ELEC_01  = 0x7844,
+	MSD_SE_MV04_DAMAGE_ELEC_02  = 0x7845,
+	MSD_SE_MV05_DAMAGE_FIRE_01  = 0x7849,
+	MSD_SE_MV05_DAMAGE_FIRE_02  = 0x784A,
+	MSD_SE_MV06_DAMAGE_WATER_01 = 0x784F,
+	MSD_SE_MV06_DAMAGE_WATER_02 = 0x7850,
+	MSD_SE_MV06_DAMAGE_WATER_03 = 0x7851,
+	MSD_SE_MV07_DAMAGE_REACT_01 = 0x7852,
+	MSD_SE_MV07_DAMAGE_REACT_02 = 0x7853,
+	MSD_SE_MV07_DAMAGE_REACT_03 = 0x7854,
+	MSD_SE_MV07_DAMAGE_REACT_04 = 0x7855,
+	MSD_SE_MV08A_DOWN_01        = 0x785D,
+	MSD_SE_MV08A_DOWN_02        = 0x785E,
+	MSD_SE_MV08A_DOWN_03        = 0x785F,
+	MSD_SE_MV08B_DOWN_REACT_01  = 0x7861,
+	MSD_SE_MV08B_DOWN_REACT_02  = 0x7862,
+	MSD_SE_MV08B_DOWN_REACT_03  = 0x7863,
+	MSD_SE_MV08B_DOWN_REACT_04  = 0x7864,
+	MSD_SE_MV10A_CRY_SHORT_01   = 0x7865,
+	MSD_SE_MV10A_CRY_SHORT_02   = 0x7866,
+	MSD_SE_MV10A_CRY_SHORT_03   = 0x7867,
+	MSD_SE_MV10B_CRY_JUMP_01    = 0x786B,
+	MSD_SE_MV12_REACT_01        = 0x7881,
+	MSD_SE_MV12_REACT_02        = 0x7882,
+	MSD_SE_MV12_REACT_03        = 0x7883,
+	MSD_SE_MV13_ACTION_SMALL_01 = 0x7884,
+	MSD_SE_MV13_ACTION_SMALL_02 = 0x7885,
+	MSD_SE_MV13_ACTION_SMALL_03 = 0x7886,
+	MSD_SE_MV15_EXERT_INST_01   = 0x788F,
+	MSD_SE_MV15_EXERT_INST_02   = 0x7890,
+	MSD_SE_MV15_REACT_SMALL_01  = 0x7891,
+	MSD_SE_MV15_REACT_SMALL_02  = 0x7892,
+	MSD_SE_MV15_REACT_SMALL_03  = 0x7893,
+	MSD_SE_MV16_EXERT_CONT_01   = 0x7094,
+	MSD_SE_MV17_EXERT_REACT_01  = 0x7899,
+	MSD_SE_MV17_EXERT_REACT_02  = 0x789A,
+	MSD_SE_MV18_FALL_AFTER_01   = 0x789E,
+	MSD_SE_MV18_FALL_AFTER_02   = 0x789F,
+	MSD_SE_MV19_JMP_PT_01       = 0x78A3,
+	MSD_SE_MV19_JMP_PT_02       = 0x78A4,
+	MSD_SE_MV20_JMP_PT_REACT_01 = 0x78A8,
+	MSD_SE_MV21_JUMP_SMALL_01   = 0x78AB,
+	MSD_SE_MV21_JUMP_SMALL_02   = 0x78AC,
+	MSD_SE_MV21_JUMP_SMALL_03   = 0x78AD,
+	MSD_SE_MV21_JUMP_SMALL_04   = 0x78AE,
+	MSD_SE_MV21_JUMP_SMALL_05   = 0x78AF,
+	MSD_SE_MV22_JUMP_MID_01     = 0x78B1,
+	MSD_SE_MV23_JUMP_LARGE_01   = 0x78B6,
+	MSD_SE_MV23_JUMP_LARGE_02   = 0x78B7,
+	MSD_SE_MV23_JUMP_LARGE_03   = 0x78B8,
+	MSD_SE_MV24_JUMP_SPECIAL_01 = 0x78B9,
+	MSD_SE_MV24_JUMP_SPECIAL_02 = 0x78BA,
+	MSD_SE_MV24_JUMP_SPECIAL_03 = 0x78BB,
+	MSD_SE_MV24_JUMP_SPECIAL_04 = 0x78BC,
+	MSD_SE_MV24_JUMP_SPECIAL_05 = 0x78BD,
+	MSD_SE_MV24_JUMP_SPECIAL_06 = 0x78BE,
+	MSD_SE_MV25A_JUMP_HUGE_01   = 0x78BF,
+	MSD_SE_MV25A_JUMP_HUGE_02   = 0x78C0,
+	MSD_SE_MV25A_WIRE_ROLL_01A  = 0x78C1,
+	MSD_SE_MV25A_WIRE_ROLL_01B  = 0x78C2,
+	MSD_SE_MV25A_WIRE_ROLL_01C  = 0x78C3,
+	MSD_SE_MV25A_WIRE_ROLL_02A  = 0x78C4,
+	MSD_SE_MV25A_WIRE_ROLL_02B  = 0x78C5,
+	MSD_SE_MV25A_WIRE_ROLL_02C  = 0x78C6,
+	MSD_SE_MV26_JUMP_REACT_01   = 0x78C7,
+	MSD_SE_MV26_JUMP_REACT_02   = 0x78C8,
+	MSD_SE_MV26_JUMP_REACT_03   = 0x78C9,
+	MSD_SE_MV26_JUMP_REACT_04   = 0x78CA,
+	MSD_SE_MV26_JUMP_REACT_05   = 0x78CB,
+	MSD_SE_MV26_JUMP_REACT_06   = 0x78CC,
+	MSD_SE_MV26_JUMP_REACT_07   = 0x78CD,
+	MSD_SE_MV26_JUMP_REACT_08   = 0x78CE,
+	MSD_SE_MV27_SPRISE_01       = 0x78CF,
+	MSD_SE_MV27_SPRISE_02       = 0x78D0,
+	MSD_SE_MV28_SPRISE_SMALL_01 = 0x78D3,
+	MSD_SE_MV28_SPRISE_SMALL_02 = 0x78D4,
+	MSD_SE_MV29_SPRISE_REACT_01 = 0x78D9,
+	MSD_SE_MV29_SPRISE_REACT_02 = 0x78DA,
+	MSD_SE_MV29_SPRISE_REACT_03 = 0x78DB,
+	MSD_SE_MV30_FRIGHT_01       = 0x78E0,
+	MSD_SE_MV30_FRIGHT_02       = 0x78E1,
+	MSD_SE_MV31_OPEN_DOOR_01    = 0x78E5,
+	MSD_SE_MV31_OPEN_DOOR_02    = 0x78E6,
+	MSD_SE_MV31_OPEN_DOOR_03    = 0x78E7,
+	MSD_SE_MV31_OPEN_DOOR_04    = 0x78E8,
+	MSD_SE_MV33_T_01            = 0x78EE,
+	MSD_SE_MV33_T_02            = 0x78EF,
+	MSD_SE_MV33_T_03            = 0x78F0,
+	MSD_SE_MV33_T_04            = 0x78F1,
+	MSD_SE_MV33_T_05            = 0x78F2,
+	MSD_SE_MV33_T_06            = 0x78F3,
+	MSD_SE_MV34_ACTION_T_01     = 0x78FB,
+	MSD_SE_MV34_ACTION_T_02     = 0x78FC,
+	MSD_SE_MV34_ACTION_T_03     = 0x78FD,
+	MSD_SE_MV38_EXERT_INST_T_01 = 0x78FE,
+	MSD_SE_MV41_JUMP_T_01       = 0x7901,
+	MSD_SE_MV41_JUMP_T_02       = 0x7902,
+	MSD_SE_MV41_JUMP_T_03       = 0x7903,
+	MSD_SE_MV42_JUMP_HUGE_T_01  = 0x7906,
+	MSD_SE_MV42_JUMP_HUGE_T_02  = 0x7907,
+	MSD_SE_MV46A_OPEN_DOOR_T_01 = 0x790B,
+	MSD_SE_MV46A_OPEN_DOOR_T_02 = 0x790C,
+	MSD_SE_MV46B_NEXT_STG_T_01  = 0x790E,
+	MSD_SE_MV46B_NEXT_STG_T_02  = 0x790F,
+	MSD_SE_MV46B_NEXT_STG_T_03  = 0x7910,
+	MSD_SE_MV_UNTEI_01          = 0x7911,
+	MSD_SE_MV_UNTEI_02          = 0x7912,
+	MSD_SE_MV_UNTEI_03          = 0x7913,
+	MSD_SE_MV_CHAO              = 0x7914,
+	MSD_SE_MV_COIN              = 0x7915,
+	MSD_SE_MV_IBIKI             = 0x7916,
+	MSD_SE_MV_IBIKI2            = 0x7917,
+	MSD_SE_YV_DAMAGE            = 0x7918,
+	MSD_SE_YV_YOSHI1            = 0x7919,
+	MSD_SE_YV_YOSHI2            = 0x791A,
+	MSD_SE_YV_YOSHI_SHORT       = 0x791B,
+	MSD_SE_YV_SURPRISED         = 0x791C,
+	MSD_SE_YV_JUMP1             = 0x791D,
+	MSD_SE_YV_JUMP2             = 0x791E,
+	MSD_SE_YV_JUMP_PYON         = 0x791F,
+	MSD_SE_YV_PERON             = 0x7920,
+	MSD_SE_YV_DELICIOUS         = 0x7921,
+	MSD_SE_YV_DELICIOUS_YOSHI1  = 0x7922,
+	MSD_SE_YV_DELICIOUS_YOSHI2  = 0x7923,
+	MSD_SE_YV_PURU_PURU         = 0x7924,
+	MSD_SE_YV_ARE               = 0x7925,
+	MSD_SE_YV_FUNBARI           = 0x7926,
+	MSD_SE_YV_FUNBARI2          = 0x7927,
+	MSD_SE_YV_FUNBARI3          = 0x7928,
+	MSD_SE_YV_WATER             = 0x7129,
+	MSD_SE_YV_PURU_PURU2        = 0x792A,
+	MSD_SE_YV_PURU_POWERFUL     = 0x792B,
+	MSD_SE_MV_DOOR              = 0x792C,
+	MSD_SE_MV_KAGEMARIO_OPENING = 0x792D,
+	MSD_SE_MV_BYEBYE            = 0x792E,
+	MSD_SE_EN_NAMEKURI_DOWN     = 0x2800,
+	MSD_SE_BOSS_WALK            = 0x2801,
+	MSD_SE_EN_NAMEKURI_DOWN_WT  = 0x2802,
+	MSD_SE_EN_COMMON_DEAD       = 0x2803,
+	MSD_SE_EN_HAMUKURI_CRUSHED  = 0x2804,
+	MSD_SE_EN_HAMUKURI_FLY      = 0x2805,
+	MSD_SE_EN_NAMEKURI_WALK     = 0x2006,
+	MSD_SE_BS_HINO_SEED_LQ_LEV  = 0x2007,
+	MSD_SE_EN_PAKKUN_SHOOT_VO   = 0x2808,
+	MSD_SE_EN_PAKKUN_SHOOT_IMI  = 0x2809,
+	MSD_SE_EN_PAKKUN_DAMAGE_VO  = 0x280A,
+	MSD_SE_EN_PAKKUN_DOWN_VO    = 0x280B,
+	MSD_SE_EN_PAKKUN_SET        = 0x280C,
+	MSD_SE_EN_PAKKUN_HIDE       = 0x280D,
+	MSD_SE_EN_HINO_EXC_ZAKO     = 0x280E,
+	MSD_SE_BS_DAMAGE_SYMBL      = 0x280F,
+	MSD_SE_EN_HMKRI_OVER        = 0x2810,
+	MSD_SE_EN_HMKRI_UP          = 0x2811,
+	MSD_SE_EN_GESO_GERO_FLY     = 0x2012,
+	MSD_SE_BS_BSPAKU_SWALLOW    = 0x2813,
+	MSD_SE_BS_BSPAKU_DRUNK      = 0x2814,
+	MSD_SE_BS_BSPAKU_GETUP1     = 0x2815,
+	MSD_SE_BS_BSPAKU_GETUP2     = 0x2816,
+	MSD_SE_BS_BSPAKU_FALL       = 0x2817,
+	MSD_SE_EN_COMMON_BOYOYON    = 0x2818,
+	MSD_SE_BS_HINO_DIVE_IMI1    = 0x2819,
+	MSD_SE_BS_HINO_DIVE_IMI2    = 0x281A,
+	MSD_SE_BS_HINO_DIVE_IMI3    = 0x281B,
+	MSD_SE_BS_HINO_DIVE_IMI4    = 0x281C,
+	MSD_SE_BS_HINO_SQURT_REAL   = 0x281D,
+	MSD_SE_BS_HINO_SQURT_IMI    = 0x281E,
+	MSD_SE_BS_DOWN_SYMBL        = 0x281F,
+	MSD_SE_BS_HINO_DOWN_FURA    = 0x2820,
+	MSD_SE_BS_HINO_DOWN_KOTE    = 0x2821,
+	MSD_SE_EN_GEN_WIND1         = 0x2822,
+	MSD_SE_EN_GEN_WIND2         = 0x2823,
+	MSD_SE_BS_HINO_RETURN_IMI   = 0x2824,
+	MSD_SE_BS_HINO_APPEAR_IMI   = 0x2825,
+	MSD_SE_EN_HMKRI_VO_ATTACK   = 0x2826,
+	MSD_SE_BS_HINO_APPEAR_WIND1 = 0x2827,
+	MSD_SE_BS_HINO_APPEAR_WIND2 = 0x2828,
+	MSD_SE_BS_HINO_APPEAR_WIND3 = 0x2829,
+	MSD_SE_BS_HINO_APPEAR_TIMP  = 0x282A,
+	MSD_SE_EN_COMMON_FLY        = 0x282B,
+	MSD_SE_EN_GESO_VO_GERO_A    = 0x282C,
+	MSD_SE_EN_GESO_VO_GERO_B    = 0x282D,
+	MSD_SE_EN_GESO_VO_DAMAGE    = 0x282E,
+	MSD_SE_EN_GESO_VO_DMG_END   = 0x282F,
+	MSD_SE_EN_GESO_VO_DOWN      = 0x2830,
+	MSD_SE_EN_GESO_VO_SET       = 0x2831,
+	MSD_SE_EN_GESO_SET_PITCH    = 0x2832,
+	MSD_SE_EN_GESO_VO_PUNCH     = 0x2833,
+	MSD_SE_EN_GESO_JUMP         = 0x2834,
+	MSD_SE_EN_GESO_LAND         = 0x2835,
+	MSD_SE_EN_GESO_WALK         = 0x2836,
+	MSD_SE_EN_HMKRI_VO_APPEAR   = 0x2837,
+	MSD_SE_EN_HMKRI_VO_DOWN     = 0x2838,
+	MSD_SE_EN_HMKRI_WALK        = 0x2839,
+	MSD_SE_EN_MAMEGESO_MELT     = 0x283A,
+	MSD_SE_EN_MAMEGESO_DOWN     = 0x283B,
+	MSD_SE_EN_COMMON_CRUSHED    = 0x283C,
+	MSD_SE_EN_NOKO_PITIPITI2    = 0x283D,
+	MSD_SE_EN_NOKO_CATCH        = 0x283E,
+	MSD_SE_EN_PUKU_PITIPITI     = 0x283F,
+	MSD_SE_BS_BSPAKU_SWING      = 0x2840,
+	MSD_SE_BS_BSPAKU_POLLUT_GND = 0x2841,
+	MSD_SE_BS_BSPAKU_FLY        = 0x2842,
+	MSD_SE_BS_BSPAKU_DOWN_END   = 0x2843,
+	MSD_SE_BS_KRPAKU_GND        = 0x2844,
+	MSD_SE_BS_BSPAKU_POLLUT_FLY = 0x2045,
+	MSD_SE_BS_BSPAKU_VO_KOMIAGE = 0x2846,
+	MSD_SE_BS_BSPAKU_VO_HAKU    = 0x2847,
+	MSD_SE_BS_BSPAKU_V_HEADEND  = 0x2848,
+	MSD_SE_BS_BSPAKU_V_RETURN_H = 0x2849,
+	MSD_SE_BS_BSPAKU_FLY_B      = 0x284A,
+	MSD_SE_BS_BSPAKU_V_LAND     = 0x284B,
+	MSD_SE_BS_BSPAKU_HIT        = 0x284C,
+	MSD_SE_BS_BSPAKU_DAMAGE     = 0x284D,
+	MSD_SE_BS_BSPAKU_DOWN       = 0x284E,
+	MSD_SE_BS_BSPAKU_RETURN     = 0x284F,
+	MSD_SE_BS_BSPAKU_FLY_NVOICE = 0x2850,
+	MSD_SE_BS_BSPAKU_FLY_NV_B   = 0x2851,
+	MSD_SE_BS_BSPAKU_POLLUT_IMI = 0x2052,
+	MSD_SE_BS_KRPAKU_GND_SMALL  = 0x2853,
+	MSD_SE_BS_KRPAKU_ROLL       = 0x2054,
+	MSD_SE_BS_BSPAKU_HIT_DOWN   = 0x2855,
+	MSD_SE_BS_KRPAKU_DOWN       = 0x2856,
+	MSD_SE_EN_BOMBHEI_LAND      = 0x2857,
+	MSD_SE_EN_BOMBHEI_BOMB      = 0x2858,
+	MSD_SE_EN_BOMBHEI_COUNT     = 0x2859,
+	MSD_SE_EN_KILLER_BOMB       = 0x285A,
+	MSD_SE_EN_MAMEGESO_FRWATER  = 0x285B,
+	MSD_SE_EN_MAMEGESO_TOWATER  = 0x285C,
+	MSD_SE_EN_KILLER_FIRE       = 0x285D,
+	MSD_SE_EN_BOMBHEI_SPARK     = 0x205E,
+	MSD_SE_EN_POPO_JUMPA        = 0x285F,
+	MSD_SE_EN_POPO_JUMPB        = 0x2860,
+	MSD_SE_EN_POPO_STUCK        = 0x2861,
+	MSD_SE_EN_TYUHANA_WALK      = 0x2862,
+	MSD_SE_EN_GESO_VO_GERO_C    = 0x2863,
+	MSD_SE_EN_POIHANA_WALK_A    = 0x2864,
+	MSD_SE_EN_POIHANA_WALK_B    = 0x2865,
+	MSD_SE_EN_SAMBOHEAD_APPEAR  = 0x2866,
+	MSD_SE_EN_GESO_GERO_LAND    = 0x2867,
+	MSD_SE_EN_BOMBHEI_ZENMAI    = 0x2068,
+	MSD_SE_BS_GESO_DOWN         = 0x2869,
+	MSD_SE_BS_GESO_DOWN_DIVE    = 0x286A,
+	MSD_SE_BS_GESO_TAKEN_HAND   = 0x206B,
+	MSD_SE_EN_TOBIPUKU_FRWATER  = 0x286C,
+	MSD_SE_EN_TOBIPUKU_TOWATER  = 0x286D,
+	MSD_SE_BS_GESO_ATK          = 0x286E,
+	MSD_SE_BS_GESO_ATK_IMPACT   = 0x286F,
+	MSD_SE_BS_GESO_LAND         = 0x2870,
+	MSD_SE_BS_GESO_GERO_LAND    = 0x2871,
+	MSD_SE_BS_GESO_HAND_OFF     = 0x2872,
+	MSD_SE_BS_GESO_RECOVER      = 0x2873,
+	MSD_SE_BS_GESO_OSEN         = 0x2874,
+	MSD_SE_EN_TYUHANA_V_JUMP    = 0x2875,
+	MSD_SE_BS_GESO_CANNON       = 0x2876,
+	MSD_SE_EN_TYUHANA_V_CHANCE  = 0x2877,
+	MSD_SE_EN_TYUHANA_V_RECOVER = 0x2878,
+	MSD_SE_EN_TYUHANA_END       = 0x2879,
+	MSD_SE_EN_TOBIPUKU_PITIPITI = 0x287A,
+	MSD_SE_EN_TYUHANA_V_PUSH    = 0x287B,
+	MSD_SE_EN_SAMBOH_JUMP_END   = 0x287C,
+	MSD_SE_EN_SAMBOH_JUMP_START = 0x287D,
+	MSD_SE_EN_PAKKUN_SEED_SINK  = 0x287E,
+	MSD_SE_EN_PAKKUN_RIP        = 0x287F,
+	MSD_SE_EN_BOMBHEI_SPARK_END = 0x2880,
+	MSD_SE_EN_POIHANA_WATER_HIT = 0x2881,
+	MSD_SE_EN_POIHANA_THROW     = 0x2882,
+	MSD_SE_EN_TYUHANA_JUMP      = 0x2883,
+	MSD_SE_BS_HANA_TUMBLE       = 0x2884,
+	MSD_SE_BS_GESO_MOUTH_HIT    = 0x2885,
+	MSD_SE_BS_GESO_WAIT1        = 0x2086,
+	MSD_SE_BS_GESO_WAIT2        = 0x2087,
+	MSD_SE_BS_GESO_V_ANGER      = 0x2888,
+	MSD_SE_BS_HANA_STEAM        = 0x2889,
+	MSD_SE_BS_HANA_WALK_RUN     = 0x288A,
+	MSD_SE_BS_GESO_BATAN        = 0x288B,
+	MSD_SE_EN_TYUHANA_SLIDE     = 0x288C,
+	MSD_SE_BS_GESO_PULL         = 0x208D,
+	MSD_SE_EN_GATEKEEPER_APPEAR = 0x288E,
+	MSD_SE_EN_GATEKEEPER_WAIT   = 0x208F,
+	MSD_SE_EN_GATEKEEPER_HIDE   = 0x2890,
+	MSD_SE_EN_GATEKEEPER_HIT    = 0x2891,
+	MSD_SE_EN_GATEKEEPER_DOWN   = 0x2892,
+	MSD_SE_DM_REFLECTION_1      = 0x2893,
+	MSD_SE_DM_REFLECTION_2      = 0x2894,
+	MSD_SE_DM_REFLECTION_3      = 0x2895,
+	MSD_SE_BS_GESO_SWING        = 0x2896,
+	MSD_SE_BS_GESO_V_CONFUSION  = 0x2897,
+	MSD_SE_BS_HANA_V_DAMAGE     = 0x2898,
+	MSD_SE_OBJ_SANDBUD_NORMAL   = 0x2099,
+	MSD_SE_OBJ_SANDBOMB_WATER_1 = 0x289A,
+	MSD_SE_OBJ_SANDBOMB_WATER_2 = 0x289B,
+	MSD_SE_OBJ_SANDBOMB_WATER_3 = 0x289C,
+	MSD_SE_OBJ_SANDBOMB_WATER_4 = 0x289D,
+	MSD_SE_OBJ_MIRROR_MOVE      = 0x209E,
+	MSD_SE_BS_HANA_BOYOYON      = 0x289F,
+	MSD_SE_BS_HANA_WRONG        = 0x28A0,
+	MSD_SE_BS_HANA_DAMAGE       = 0x28A1,
+	MSD_SE_BS_GESO_V_UPSET      = 0x28A2,
+	MSD_SE_EN_TYUHANA_FLY       = 0x28A3,
+	MSD_SE_OBJ_SANDBOMB_BANG    = 0x28A4,
+	MSD_XX_DM_OSEN_DISAPPEAR    = 0x28A5,
+	MSD_SE_EN_COMMON_TWINKLE    = 0x28A6,
+	MSD_SE_EN_NAMEKURI_JUMP     = 0x28A7,
+	MSD_SE_MA_SLIP_TEMP         = 0x20A8,
+	MSD_SE_EN_KILLER_FLY        = 0x20A9,
+	MSD_SE_BS_WANWAN_HOP        = 0x28AA,
+	MSD_SE_EN_WANWAN_FLAME      = 0x20AB,
+	MSD_SE_EN_WANWAN_HOP        = 0x28AC,
+	MSD_SE_EN_IGAIGA_BOUND      = 0x28AD,
+	MSD_SE_EN_TYOROBE_V_APPEAR  = 0x28AE,
+	MSD_SE_EN_TYOROBE_V_THROW   = 0x28AF,
+	MSD_SE_EN_TYOROBE_V_HIDE    = 0x28B0,
+	MSD_SE_EN_TYOROBE_V_DAMAGE1 = 0x28B1,
+	MSD_SE_EN_TYOROBE_V_DAMAGE2 = 0x28B2,
+	MSD_SE_EN_TYOROBE_V_DOWN    = 0x28B3,
+	MSD_SE_EN_KAZEKUN_APPEAR    = 0x28B4,
+	MSD_SE_EN_KAZEKUN_WAIT      = 0x20B5,
+	MSD_SE_EN_KAZEKUN_READY     = 0x28B6,
+	MSD_SE_EN_KAZEKUN_ATTACK    = 0x28B7,
+	MSD_SE_BS_BSPAKU_SLAP       = 0x28B8,
+	MSD_SE_EN_WANWAN_BARK       = 0x28B9,
+	MSD_SE_BS_WANWAN_BARK       = 0x28BA,
+	MSD_SE_EN_DENNOKO_SPARK1    = 0x20BB,
+	MSD_SE_EN_DENNOKO_SPARK2    = 0x20BC,
+	MSD_SE_EN_WANWAN_COOL       = 0x28BD,
+	MSD_SE_BS_WANWAN_COOL       = 0x28BE,
+	MSD_SE_EN_WIRETRAP_WAIT     = 0x20BF,
+	MSD_SE_BS_WANWAN_LOCK       = 0x28C0,
+	MSD_SE_EN_DENNOKO_KOURA_ON  = 0x28C1,
+	MSD_SE_EN_POPO_GROW         = 0x20C2,
+	MSD_SE_EN_MOEKURI_FLAME     = 0x20C3,
+	MSD_SE_EN_WANWAN_TO_COOL    = 0x28C4,
+	MSD_SE_BS_WANWAN_TO_COOL    = 0x28C5,
+	MSD_SE_OBJ_SAMDBOMB_REVERSE = 0x20C6,
+	MSD_SE_EN_CANNON_OPEN       = 0x28C7,
+	MSD_SE_EN_CANNON_MOVE       = 0x20C8,
+	MSD_SE_EN_DENNOKO_KOURA_OFF = 0x28C9,
+	MSD_SE_EN_TELESA_FLY        = 0x28CA,
+	MSD_SE_EN_TELESA_FIX        = 0x28CB,
+	MSD_SE_EN_TELESA_CRUSH      = 0x28CC,
+	MSD_SE_EN_POPO_POP          = 0x28CD,
+	MSD_SE_EN_POPO_FLY          = 0x20CE,
+	MSD_SE_EN_TAMANOKO_YARN1    = 0x28CF,
+	MSD_SE_EN_TAMANOKO_YARN2    = 0x28D0,
+	MSD_SE_BS_WANWAN_COOL_MORE  = 0x28D1,
+	MSD_SE_BS_WANWAN_M_DRAG     = 0x20D2,
+	MSD_SE_BS_TELESA_V_APPEAR   = 0x28D3,
+	MSD_SE_BS_TELESA_V_LAUGH1   = 0x28D4,
+	MSD_SE_BS_TELESA_DISAPPEAR  = 0x28D5,
+	MSD_SE_BS_TELESA_SPICY_HIT  = 0x28D6,
+	MSD_SE_BS_TELESA_BREATHE    = 0x28D7,
+	MSD_SE_EN_TELESA_APPEAR     = 0x28D8,
+	MSD_SE_BS_TELESA_V_LAUGH2   = 0x28D9,
+	MSD_SE_BS_TELESA_LICK       = 0x28DA,
+	MSD_SE_BS_TELESA_DOWN       = 0x28DB,
+	MSD_SE_BS_TELESA_V_LAUGH1B  = 0x28DC,
+	MSD_SE_EN_WANWAN_BARK2      = 0x20DD,
+	MSD_SE_EN_WANWAN_HOLD       = 0x28DE,
+	MSD_SE_EN_WANWAN_PULL       = 0x20DF,
+	MSD_SE_EN_WANWAN_THROWN     = 0x28E0,
+	MSD_SE_EN_WANWAN_REFLECT    = 0x28E1,
+	MSD_SE_EN_WANWAN_RECOVER    = 0x28E2,
+	MSD_SE_EN_WANWAN_DOWN       = 0x28E3,
+	MSD_SE_EN_WANWAN_AFTER      = 0x20E4,
+	MSD_SE_BS_BSPAKU_SWING2     = 0x28E5,
+	MSD_SE_BS_HANA_DOWN         = 0x28E6,
+	MSD_SE_BS_TELESA_THANKYOU   = 0x28E7,
+	MSD_SE_EN_SAMBO_DOWN        = 0x28E8,
+	MSD_SE_EN_TAMANOKO_WALK     = 0x28E9,
+	MSD_SE_EN_TAMANOKO_CREEP    = 0x28EA,
+	MSD_SE_EN_AMINOKO_WALK1     = 0x28EB,
+	MSD_SE_EN_AMINOKO_WALK2     = 0x28EC,
+	MSD_SE_EN_DENNOKO_ELEC_DOWN = 0x28ED,
+	MSD_SE_EN_DENNOKO_JITABATA  = 0x28EE,
+	MSD_SE_EN_DENNOKO_SHOCKED   = 0x28EF,
+	MSD_SE_EN_DENNOKO_WALK      = 0x28F0,
+	MSD_SE_EN_AMINOKO_SPARK     = 0x20F1,
+	MSD_SE_EN_SAMBOH_HIDE       = 0x28F2,
+	MSD_SE_EN_TORI_SHIT         = 0x28F3,
+	MSD_SE_EN_BEENEST_OFF       = 0x28F4,
+	MSD_SE_EN_BEENEST_FALL      = 0x28F5,
+	MSD_SE_EN_BEENEST_LAND      = 0x28F6,
+	MSD_SE_EN_BEENEST_SWING     = 0x28F7,
+	MSD_SE_EN_KOHANA_WAKEUP2    = 0x28F8,
+	MSD_SE_EN_TAMANOKO_DROPOK   = 0x28F9,
+	MSD_SE_EN_KOHANA_GETUP      = 0x28FA,
+	MSD_SE_EN_NAMEKURI_DOWN2    = 0x28FB,
+	MSD_XX_EN_CANNON_LAND       = 0x28FC,
+	MSD_SE_BS_TELESA_DAMAGE     = 0x28FD,
+	MSD_SE_BS_TELESA_FIRE       = 0x20FE,
+	MSD_SE_EN_KILLER_FLY_KUPPA  = 0x20FF,
+	MSD_SE_EN_TORI_FLY          = 0x2900,
+	MSD_SE_BS_HANA_V_TUMBLE     = 0x2901,
+	MSD_SE_BS_GESO_VO_ATTACK    = 0x2902,
+	MSD_SE_EN_MOEKURI_COOL      = 0x2903,
+	MSD_SE_EN_TAMANOKO_SHOOK    = 0x2904,
+	MSD_SE_EN_POIHANA_WATERHIT2 = 0x2905,
+	MSD_SE_EN_BEE_GROUP         = 0x2106,
+	MSD_SE_EN_BEE_3             = 0x2107,
+	MSD_SE_EN_BEE_2             = 0x2108,
+	MSD_SE_EN_BEE_1             = 0x2109,
+	MSD_SE_BS_BSPAKU_GARGLE     = 0x210A,
+	MSD_SE_EN_GATEKEEPER_BASE   = 0x210B,
+	MSD_SE_BS_BSPAKU_TORNADO    = 0x210C,
+	MSD_SE_EN_POIHANA_TO_SLEEP  = 0x290D,
+	MSD_SE_EN_KUMO_WALK_WALL_A  = 0x290E,
+	MSD_SE_EN_WANWAN_1ST_WATER  = 0x290F,
+	MSD_XX_EN_COMMON_W_HIT_OK   = 0x2910,
+	MSD_SE_EN_TORI_SHIT_LAND    = 0x2911,
+	MSD_SE_BS_GESO_WATER_HIT    = 0x2912,
+	MSD_SE_BS_KRPAKU_SINK       = 0x2913,
+	MSD_XX_EN_CANNON_DOWN       = 0x2914,
+	MSD_SE_BS_WANWAN_REBURN1    = 0x2915,
+	MSD_SE_BS_WANWAN_REBURN2    = 0x2916,
+	MSD_SE_BS_WANWAN_DIVE       = 0x2917,
+	MSD_SE_EN_TAMANOKO_JUMP     = 0x2918,
+	MSD_SE_BS_BSPAKU_SHOOT      = 0x2919,
+	MSD_SE_EN_SAMBO_WAIT        = 0x291A,
+	MSD_SE_EN_SAMBO_ATTACK      = 0x291B,
+	MSD_SE_EN_KILLER_SEARCH     = 0x291C,
+	MSD_SE_EN_KAZEKUN_DOWN      = 0x291D,
+	MSD_SE_EN_TORI_WATER_HIT    = 0x291E,
+	MSD_SE_EN_TORI_DOWN         = 0x291F,
+	MSD_SE_EN_AMIKING_FLY       = 0x2120,
+	MSD_SE_EN_AMIKING_DIVE      = 0x2921,
+	MSD_SE_EN_TOBIPUKU_BITE     = 0x2922,
+	MSD_SE_EN_TOBIPUKU_CHEW     = 0x2123,
+	MSD_SE_BS_TELESA_RLT_STOP   = 0x2924,
+	MSD_SE_BS_TELESA_RLT_DOWN   = 0x2125,
+	MSD_SE_BS_TELESA_RLT_SET    = 0x2926,
+	MSD_XX_BS_TELESA_RLT_MOVE3  = 0x2127,
+	MSD_XX_BS_TELESA_RLT_MOVE2  = 0x2128,
+	MSD_XX_BS_TELESA_RLT_MOVE1  = 0x2129,
+	MSD_SE_BS_TELESA_SLT_LAND   = 0x292A,
+	MSD_SE_BS_TELESA_SLT_SPIN   = 0x212B,
+	MSD_SE_BS_TELESA_SLT_STOP   = 0x292C,
+	MSD_SE_EN_TAMANOKO_BORN     = 0x292D,
+	MSD_SE_EN_IGAIGA_FIRE       = 0x292E,
+	MSD_SE_EN_IGAIGA_ROLL       = 0x212F,
+	MSD_SE_EN_IGAIGA_GROW       = 0x2130,
+	MSD_SE_EN_IGAIGA_DOWN1      = 0x2931,
+	MSD_SE_EN_IGAIGA_DOWN2      = 0x2932,
+	MSD_SE_EN_IGAIGA_WATER_DOWN = 0x2933,
+	MSD_SE_EN_GENERATOR_BASE    = 0x2134,
+	MSD_SE_EN_DENNOKO_VOICE1    = 0x2935,
+	MSD_SE_EN_AMINOKO_FLY       = 0x2936,
+	MSD_SE_EN_KM_TELSA_REVEAL   = 0x2937,
+	MSD_SE_EN_DB_TELSA_EATEN    = 0x2938,
+	MSD_SE_EN_DOROKURI_JUMP     = 0x2939,
+	MSD_SE_EN_SAMBOHEAD_DOWN    = 0x293A,
+	MSD_SE_EN_SAMBOHEAD_CRUSH   = 0x293B,
+	MSD_SE_BS_BSPAKU_FALL_END   = 0x293C,
+	MSD_SE_EN_TELSA_RECOVER     = 0x293D,
+	MSD_SE_EN_GENERATOR_DOWN    = 0x293E,
+	MSD_SE_BS_TELESA_FANFALE_1  = 0x293F,
+	MSD_SE_BS_TELESA_FANFALE_2  = 0x2940,
+	MSD_SE_EN_YUMBO_TURN        = 0x2141,
+	MSD_SE_EN_TAMANOKO_DROP_NG  = 0x2942,
+	MSD_SE_EN_TELESA_DISAPPEAR  = 0x2943,
+	MSD_SE_BS_TELESA_FRUIT_HIT  = 0x2944,
+	MSD_SE_BS_TELESA_V_ROLL     = 0x2945,
+	MSD_SE_EN_TOBIPUKU_SWIM     = 0x2946,
+	MSD_SE_EN_TOBIPUKU_SEARCH   = 0x2947,
+	MSD_SE_EN_AMINOKO_WALK3     = 0x2949,
+	MSD_SE_EN_AMINOKO_WALK4     = 0x294A,
+	MSD_SE_BS_BSPAKU_SNORE_INH  = 0x294B,
+	MSD_SE_BS_BSPAKU_SNORE_EXH  = 0x294C,
+	MSD_SE_BS_TELESA_FANFALE_3  = 0x294D,
+	MSD_SE_EN_CANNON_FIRE_MARIO = 0x294E,
+	MSD_SE_EN_AMIKING_SPARK     = 0x214F,
+	MSD_SE_EN_AMINOKO_ANGRY     = 0x2950,
+	MSD_SE_EN_MOEKURI_REBURN    = 0x2951,
+	MSD_SE_EN_TAMANOKO_DOWN     = 0x2952,
+	MSD_SE_BS_GESO_MHIT_NOBOICE = 0x2953,
+	MSD_SE_EN_ORANGESEAL_WAIT   = 0x2154,
+	MSD_SE_EN_KOHANA_WAKEUP3    = 0x2955,
+	MSD_XX_EN_TYOROBE_BYE       = 0x2956,
+	MSD_SE_EN_TAMANOKO_BLOOM    = 0x2957,
+	MSD_SE_EN_TAMANOKO_ESCAPE   = 0x2958,
+	MSD_SE_EN_SAMBOHEAD_W_HIT   = 0x2959,
+	MSD_SE_OBJ_CORK_WEEP        = 0x295A,
+	MSD_SE_EN_TOBIPUKU_V_DOWN   = 0x295B,
+	MSD_SE_BS_GESO_ROLL         = 0x215C,
+	MSD_SE_OBJ_FALL_JET         = 0x295D,
+	MSD_SE_EN_DANGOKURI_TURN    = 0x295E,
+	MSD_SE_EN_COMMON_SMOKE      = 0x295F,
+	MSD_SE_EN_YUMBO_BLOW        = 0x2160,
+	MSD_SE_EN_MAMEGESO_W_HIT    = 0x2961,
+	MSD_SE_EN_DB_TELESA_INHALE  = 0x2962,
+	MSD_SE_EN_DB_TELESA_EXHALE  = 0x2963,
+	MSD_SE_EN_HANEKURI_FLY      = 0x2164,
+	MSD_SE_EN_HANEKURI_ATTACK   = 0x2965,
+	MSD_SE_BS_WANWAN_ESCAPE1    = 0x2966,
+	MSD_SE_BS_WANWAN_ESCAPE2    = 0x2967,
+	MSD_SE_BS_TELESA_ESCAPE     = 0x2968,
+	MSD_SE_EN_PAKKUN_SEED_FLY   = 0x2169,
+	MSD_SE_EN_MOEPUKU_TOWATER   = 0x296A,
+	MSD_XX_OBJ_CALM_FLAME       = 0x216B,
+	MSD_SE_EN_YUMBO_SINK        = 0x296C,
+	MSD_SE_DM_FLY_TO_PINNNA     = 0x296D,
+	MSD_SE_OBJ_POSTER_RIP1      = 0x296E,
+	MSD_SE_OBJ_POSTER_RIP2      = 0x296F,
+	MSD_SE_OBJ_CLOSET_OPEN      = 0x2970,
+	MSD_XX_ENV_FALL_JET_LEVEL   = 0x2171,
+	MSD_SE_TITLE_M              = 0x2972,
+	MSD_SE_M_BOSW_CHAIN_ANG_F   = 0x2973,
+	MSD_SE_M_BOSW_CHAIN_ANG_R   = 0x2974,
+	MSD_SE_M_BOSW_CHAIN_CAL_F   = 0x2975,
+	MSD_SE_M_BOSW_CHAIN_CAL_R   = 0x2976,
+	MSD_SE_M_YUNBO_DANCE_JUMP   = 0x2977,
+	MSD_SE_M_YUNBO_DANCE_SING   = 0x2978,
+	MSD_SE_OBJ_BALLOON_BREAK    = 0x2979,
+	MSD_SE_EN_TAMANOKO_BLOOM2   = 0x297A,
+	MSD_SE_EN_HANEKURI_ATTACKNG = 0x297B,
+	MSD_SE_EN_SANDBIRD_CRY      = 0x217C,
+	MSD_SE_ENV_SANDBIRD_WIND    = 0x217D,
+	MSD_XX_MA_INTO_DOKAN        = 0x297E,
+	MSD_SE_NPC_KANARIA          = 0x8800,
+	MSD_SE_NPC_KANARIA2         = 0x8801,
+	MSD_SE_NPC_MONTE_M_WALK     = 0x8802,
+	MSD_SE_NPC_MONTE_M_VOICE    = 0x8803,
+	MSD_SE_NPC_MONTE_W_VOICE    = 0x8804,
+	MSD_SE_NPC_MONTE_C_VOICE    = 0x8805,
+	MSD_SE_NPC_MONTE_W_WALK     = 0x8806,
+	MSD_SE_NPC_APPEAR_ITEM      = 0x8807,
+	MSD_SE_NPC_MONTE_C_CLEANUP  = 0x8808,
+	MSD_SE_NPC_MONTE_C_RUBBING  = 0x8809,
+	MSD_SE_NPC_MONTE_C_SWING    = 0x880A,
+	MSD_SE_NPC_MONTE_C_JUMP     = 0x880B,
+	MSD_SE_NPC_MONTE_C_PROPEL   = 0x880C,
+	MSD_SE_NPC_MONTE_C_LAND     = 0x880D,
+	MSD_SE_NPC_MONTE_C_PROPE_LV = 0x800E,
+	MSD_SE_NPC_MONTE_PRESS      = 0x880F,
+	MSD_SE_NPC_MONTEW_C_CLEANUP = 0x8810,
+	MSD_SE_NPC_MONTEW_C_RUBBING = 0x8811,
+	MSD_SE_NPC_MONTEW_C_SWING   = 0x8812,
+	MSD_SE_NPC_MONTEW_C_JUMP    = 0x8813,
+	MSD_SE_NPC_MONTEW_C_PROPEL  = 0x8814,
+	MSD_SE_NPC_MONTEW_C_LAND    = 0x8815,
+	MSD_SE_NPC_MONTEW_C_PROP_LV = 0x8016,
+	MSD_SE_NPC_BURNING          = 0x8017,
+	MSD_SE_NPC_MARE_M_WALK      = 0x8818,
+	MSD_SE_NPC_MARE_M_SWING     = 0x8819,
+	MSD_SE_NPC_MARE_M_JUMP      = 0x881A,
+	MSD_SE_NPC_MARE_M_LAND      = 0x881B,
+	MSD_SE_NPC_MARE_M_CLEANUP   = 0x881C,
+	MSD_SE_NPC_MARE_M_RUBBING   = 0x881D,
+	MSD_SE_NPC_MARE_PRESS       = 0x881E,
+	MSD_SE_NPC_MARE_HUE_RANDOM  = 0x881F,
+	MSD_SE_NPC_MARE_W_WALK      = 0x8820,
+	MSD_SE_NPC_MARE_W_SWING     = 0x8821,
+	MSD_SE_NPC_MARE_W_JUMP      = 0x8822,
+	MSD_SE_NPC_MARE_W_LAND      = 0x8823,
+	MSD_SE_NPC_MARE_W_CLEANUP   = 0x8824,
+	MSD_SE_NPC_MARE_W_RUBBING   = 0x8825,
+	MSD_SE_NPC_COMMON_PRESS     = 0x8826,
+	MSD_SE_NPC_KINOPJ_PRESS     = 0x8827,
+	MSD_SE_NPC_KINOP_WALK       = 0x8828,
+	MSD_SE_NPC_KINOP_JUMP       = 0x8829,
+	MSD_SE_NPC_KINOP_LAND       = 0x882A,
+	MSD_SE_NPC_KINOP_SPRAY      = 0x882B,
+	MSD_SE_NPC_KINOJ_WALK       = 0x882C,
+	MSD_SE_NPC_KINOJ_JUMP       = 0x882D,
+	MSD_SE_NPC_KINOJ_LAND       = 0x882E,
+	MSD_SE_NPC_POLUSION_SWING   = 0x882F,
+	MSD_SE_NPC_PEACH_WALK       = 0x8830,
+	MSD_SE_NPC_TANUKI_PON       = 0x8831,
+	MSD_SE_NPC_TANUKI_HAT       = 0x8832,
+	MSD_SE_NPC_SUNFL_LEAF       = 0x8833,
+	MSD_SE_NPC_SUNFL_NECK       = 0x8834,
+	MSD_SE_NPC_SUNFL_GROWTH     = 0x8835,
+	MSD_SE_NPC_FIRE_FIGHTING    = 0x8837,
+	MSD_SE_NPC_MONTE_C_WATER    = 0x8838,
+	MSD_SE_NPC_MONTE_C_METAL_UP = 0x8839,
+	MSD_SE_NPC_MONTE_C_METAL_DW = 0x883A,
+	MSD_SE_NPC_MONTE_C_MOP      = 0x883B,
+	MSD_SE_NPC_MONTE_C_UKULELE1 = 0x883C,
+	MSD_SE_NPC_MONTE_C_UKULELE2 = 0x883D,
+	MSD_SE_NPC_MONTE_C_UKULELE3 = 0x883E,
+	MSD_SE_NPC_MONTE_C_UKU_CUT3 = 0x883F,
+	MSD_SE_NPC_MONTE_C_UKUL_CUT = 0x8840,
+	MSD_SE_NPC_MONTE_C_UKUL_RND = 0x8841,
+	MSD_SE_NPC_MONTE_C_UKU_CUT2 = 0x8842,
+	MSD_SE_SY_SOUT_MONO         = 0x8843,
+	MSD_SE_NPC_VAS_MONTEM_PRESS = 0x8844,
+	MSD_SE_NPC_VAS_MONTEW_PRESS = 0x8845,
+	MSD_SE_NPC_VAS_MAREM_PRESS  = 0x8846,
+	MSD_SE_NPC_VAS_MAREW_PRESS  = 0x8847,
+	MSD_SE_NPC_VAS_KINOJI_PRESS = 0x8848,
+	MSD_SE_NPC_VAS_KINOPI_PRESS = 0x8849,
+	MSD_SE_SY_SOUT_STEREO       = 0x884A,
+	MSD_SE_SY_SOUT_SURROUND     = 0x884B,
+	MSD_SE_NPC_MONTE_C_UKULE2_0 = 0x884C,
+	MSD_SE_NPC_MONTE_C_UKULE2_1 = 0x884D,
+	MSD_SE_NPC_MONTE_C_UCUT2    = 0x884E,
+	MSD_SE_NPC_VM_PEACH_NORMAL  = 0x8850,
+	MSD_SE_NPC_VM_PEACH_SURPRS  = 0x8851,
+	MSD_SE_NPC_VM_PEACH_WORRY   = 0x8852,
+	MSD_SE_NPC_VM_PEACH_ANGER_L = 0x8853,
+	MSD_SE_NPC_VM_PEACH_APPEAL  = 0x8854,
+	MSD_SE_NPC_VM_PEACH_DOUBT   = 0x8855,
+	MSD_SE_NPC_VM_KINOJ_NORMAL  = 0x8856,
+	MSD_SE_NPC_VM_KINOJ_EXCIT   = 0x8857,
+	MSD_SE_NPC_VM_KINOJ_ADVICE  = 0x8858,
+	MSD_SE_NPC_VM_KINOJ_CONFUSE = 0x8859,
+	MSD_SE_NPC_VM_KINOJ_ASK     = 0x885A,
+	MSD_SE_NPC_VM_KINOP_NORMAL  = 0x885B,
+	MSD_SE_NPC_VM_KINOP_CRY_S   = 0x885C,
+	MSD_SE_NPC_VM_KINOP_CRY_L   = 0x885D,
+	MSD_SE_NPC_VM_KINOP_PEACE   = 0x885E,
+	MSD_SE_NPC_VM_KINOP_SAD_S   = 0x885F,
+	MSD_SE_NPC_VM_KINOP_ASK     = 0x8860,
+	MSD_SE_NPC_VM_KINOP_CONFUSE = 0x8861,
+	MSD_SE_NPC_VM_KINOP_RECOVER = 0x8862,
+	MSD_SE_NPC_VM_KINOP_SAD_L   = 0x8863,
+	MSD_SE_NPC_VM_MTMAN_NORMAL  = 0x8864,
+	MSD_SE_NPC_VM_MTMAN_LOST    = 0x8865,
+	MSD_SE_NPC_VM_MONTM_NORMAL  = 0x8866,
+	MSD_SE_NPC_VM_MONTM_LAUGH_D = 0x8867,
+	MSD_SE_NPC_VM_MONTM_DSIGST  = 0x8868,
+	MSD_SE_NPC_VM_MONTM_ANGER_S = 0x8869,
+	MSD_SE_NPC_VM_MONTM_SURPRS  = 0x886A,
+	MSD_SE_NPC_VM_MONTM_DOUBT   = 0x886B,
+	MSD_SE_NPC_VM_MONTM_DISPLES = 0x886C,
+	MSD_SE_NPC_VM_MONTM_CONFUSE = 0x886D,
+	MSD_SE_NPC_VM_MONTM_REGRET  = 0x886E,
+	MSD_SE_NPC_VM_MONTM_PROUD   = 0x886F,
+	MSD_SE_NPC_VM_MONTM_RECOVER = 0x8870,
+	MSD_SE_NPC_VM_MONTM_INVITN  = 0x8871,
+	MSD_SE_NPC_VM_MONTM_QUESTN  = 0x8872,
+	MSD_SE_NPC_VM_MONTM_LAUGH   = 0x8873,
+	MSD_SE_NPC_VM_MONTM_THANKS  = 0x8874,
+	MSD_SE_NPC_VM_MONTW_NORAML  = 0x8875,
+	MSD_SE_NPC_VM_MONTW_REGRET  = 0x8876,
+	MSD_SE_NPC_VM_MONTW_INVITN  = 0x8877,
+	MSD_SE_NPC_VM_MONTW_LAUGH   = 0x8878,
+	MSD_SE_NPC_VM_MAREM_NORMAL  = 0x8879,
+	MSD_SE_NPC_VM_MAREM_REGRET  = 0x887A,
+	MSD_SE_NPC_VM_MAREM_LAUGH   = 0x887B,
+	MSD_SE_NPC_VM_MAREM_APPEAL  = 0x887C,
+	MSD_SE_NPC_VM_MAREM_SURPRS  = 0x887D,
+	MSD_SE_NPC_VM_MAREM_SAD     = 0x887E,
+	MSD_SE_NPC_VM_MAREM_ASK     = 0x887F,
+	MSD_SE_NPC_VM_MAREM_PROMPT  = 0x8880,
+	MSD_SE_NPC_VM_MAREM_THANKS  = 0x8881,
+	MSD_SE_NPC_VM_MAREW_NORMAL  = 0x8882,
+	MSD_SE_NPC_VM_MAREW_REGRET  = 0x8883,
+	MSD_SE_NPC_VM_MAREW_LAUGH   = 0x8884,
+	MSD_SE_NPC_VM_MAREW_APPEAL  = 0x8885,
+	MSD_SE_NPC_VM_MAREW_SURPRS  = 0x8886,
+	MSD_SE_NPC_VM_MAREW_SAD     = 0x8887,
+	MSD_SE_NPC_VM_MAREW_ASK     = 0x8888,
+	MSD_SE_NPC_VM_MAREW_PROMPT  = 0x8889,
+	MSD_SE_NPC_VM_MAREW_THANKS  = 0x888A,
+	MSD_SE_NPC_VM_MAREJ_NORMAL  = 0x888B,
+	MSD_SE_NPC_VM_MAREJ_REGRET  = 0x888C,
+	MSD_SE_NPC_VM_MAREJ_LAUGH   = 0x888D,
+	MSD_SE_NPC_VM_MAREJ_APPEAL  = 0x888E,
+	MSD_SE_NPC_VM_MAREJ_SURPRS  = 0x888F,
+	MSD_SE_NPC_VM_MAREJ_SAD     = 0x8890,
+	MSD_SE_NPC_VM_MAREJ_ASK     = 0x8891,
+	MSD_SE_NPC_VM_MAREJ_PROMPT  = 0x8892,
+	MSD_SE_NPC_VM_MAREJ_THANKS  = 0x8893,
+	MSD_SE_NPC_VM_TANUK_NORMAL  = 0x8894,
+	MSD_SE_NPC_VM_SUNFP_JOY     = 0x8895,
+	MSD_SE_NPC_VM_SUNFP_SAD     = 0x8896,
+	MSD_SE_NPC_VM_SUNFC_JOY     = 0x8897,
+	MSD_SE_NPC_VM_SUNFC_SAD     = 0x8898,
+	MSD_SE_NPC_VM_MONTW_LAUGH_D = 0x8899,
+	MSD_SE_NPC_VM_MONTW_DSIGST  = 0x889A,
+	MSD_SE_NPC_VM_MONTW_ANGER_S = 0x889B,
+	MSD_SE_NPC_VM_MONTW_SURPRS  = 0x889C,
+	MSD_SE_NPC_VM_MONTW_DOUBT   = 0x889D,
+	MSD_SE_NPC_VM_MONTW_DISPLES = 0x889E,
+	MSD_SE_NPC_VM_MONTW_CONFUSE = 0x889F,
+	MSD_SE_NPC_VM_MONTW_PROUD   = 0x88A0,
+	MSD_SE_NPC_VM_MONTW_RECOVER = 0x88A1,
+	MSD_SE_NPC_VM_MONTW_QUESTN  = 0x88A2,
+	MSD_SE_NPC_VM_MONTW_THANKS  = 0x88A3,
+	MSD_SE_NPC_VM_MAREM_FLUTE   = 0x88A4,
+	MSD_SE_NPC_VM_SUNFP_JOY2    = 0x88A5,
+	MSD_SE_NPC_VM_FMARIO_PSHOT  = 0x88A6,
+	MSD_SE_NPC_VAS_CMONTM_PRESS = 0x88AB,
+	MSD_SE_NPC_VAS_CMONTW_PRESS = 0x88AC,
+	MSD_SE_NPC_VAS_CMAREM_PRESS = 0x88AD,
+	MSD_SE_NPC_VAS_CMAREW_PRESS = 0x88AE,
+	MSD_SE_NPC_VAS_MAREJ_PRESS  = 0x88AF,
+	MSD_SE_NPC_VM_NOZ_WARNG_L   = 0x88B0,
+	MSD_SE_NPC_VM_NOZ_WARNG_H   = 0x88B1,
+	MSD_SE_NPC_VM_NOZ_CAUTN_L   = 0x88B2,
+	MSD_SE_NPC_VM_NOZ_CAUTN_H   = 0x88B3,
+	MSD_SE_NPC_VM_NOZ_HUMOR     = 0x88B4,
+	MSD_SE_NPC_VM_NOZ_DOWN      = 0x88B5,
+	MSD_SE_NPC_VM_NOZ_WORRY_L   = 0x88B6,
+	MSD_SE_NPC_VM_NOZ_WORRY_H   = 0x88B7,
+	MSD_SE_NPC_VM_NOZ_ENCOURAGE = 0x88B8,
+	MSD_SE_NPC_VM_NOZ_CONGRATU  = 0x88B9,
+	MSD_SE_NPC_VM_NOZ_COMPLAINT = 0x88BA,
+	MSD_SE_NPC_VA_PEACH_CALL    = 0x80BC,
+	MSD_SE_NPC_VA_MONTM_TROUB2  = 0x88BD,
+	MSD_SE_NPC_VA_MONTW_ATTENT2 = 0x80BE,
+	MSD_SE_NPC_VA_PEACH_CARRY   = 0x80BF,
+	MSD_SE_NPC_VM_CMNTM_NORMAL  = 0x88C0,
+	MSD_SE_NPC_VM_CMNTM_LAUGH_D = 0x88C1,
+	MSD_SE_NPC_VM_CMNTM_DSIGST  = 0x88C2,
+	MSD_SE_NPC_VM_CMNTM_ANGER_S = 0x88C3,
+	MSD_SE_NPC_VM_CMNTM_SURPRS  = 0x88C4,
+	MSD_SE_NPC_VM_CMNTM_DOUBT   = 0x88C5,
+	MSD_SE_NPC_VM_CMNTM_DISPLES = 0x88C6,
+	MSD_SE_NPC_VM_CMNTM_CONFUSE = 0x88C7,
+	MSD_SE_NPC_VM_CMNTM_REGRET  = 0x88C8,
+	MSD_SE_NPC_VM_CMNTM_PROUD   = 0x88C9,
+	MSD_SE_NPC_VM_CMNTM_RECOVER = 0x88CA,
+	MSD_SE_NPC_VM_CMNTM_INVITN  = 0x88CB,
+	MSD_SE_NPC_VM_CMNTM_QUESTN  = 0x88CC,
+	MSD_SE_NPC_VM_CMNTM_LAUGH   = 0x88CD,
+	MSD_SE_NPC_VM_CMNTM_THANKS  = 0x88CE,
+	MSD_SE_NPC_VM_CMNTW_NORAML  = 0x88CF,
+	MSD_SE_NPC_VM_CMNTW_REGRET  = 0x88D0,
+	MSD_SE_NPC_VM_CMNTW_INVITN  = 0x88D1,
+	MSD_SE_NPC_VM_CMNTW_LAUGH   = 0x88D2,
+	MSD_SE_NPC_VM_CMAREM_NORMAL = 0x88D3,
+	MSD_SE_NPC_VM_CMAREM_REGRET = 0x88D4,
+	MSD_SE_NPC_VM_CMAREM_LAUGH  = 0x88D5,
+	MSD_SE_NPC_VM_CMAREM_APPEAL = 0x88D6,
+	MSD_SE_NPC_VM_CMAREM_SURPRS = 0x88D7,
+	MSD_SE_NPC_VM_CMAREM_SAD    = 0x88D8,
+	MSD_SE_NPC_VM_CMAREM_ASK    = 0x88D9,
+	MSD_SE_NPC_VM_CMAREM_PROMPT = 0x88DA,
+	MSD_SE_NPC_VM_CMAREM_THANKS = 0x88DB,
+	MSD_SE_NPC_VM_CMAREW_NORMAL = 0x88DC,
+	MSD_SE_NPC_VM_CMAREW_REGRET = 0x88DD,
+	MSD_SE_NPC_VM_CMAREW_LAUGH  = 0x88DE,
+	MSD_SE_NPC_VM_CMAREW_APPEAL = 0x88DF,
+	MSD_SE_NPC_VM_CMAREW_SURPRS = 0x88E0,
+	MSD_SE_NPC_VM_CMAREW_SAD    = 0x88E1,
+	MSD_SE_NPC_VM_CMAREW_ASK    = 0x88E2,
+	MSD_SE_NPC_VM_CMAREW_PROMPT = 0x88E3,
+	MSD_SE_NPC_VM_CMAREW_THANKS = 0x88E4,
+	MSD_SE_NPC_VM_CMNTW_LAUGH_D = 0x88E5,
+	MSD_SE_NPC_VM_CMNTW_DSIGST  = 0x88E6,
+	MSD_SE_NPC_VM_CMNTW_ANGER_S = 0x88E7,
+	MSD_SE_NPC_VM_CMNTW_SURPRS  = 0x88E8,
+	MSD_SE_NPC_VM_CMNTW_DOUBT   = 0x88E9,
+	MSD_SE_NPC_VM_CMNTW_DISPLES = 0x88EA,
+	MSD_SE_NPC_VM_CMNTW_CONFUSE = 0x88EB,
+	MSD_SE_NPC_VM_CMNTW_PROUD   = 0x88EC,
+	MSD_SE_NPC_VM_CMNTW_RECOVER = 0x88ED,
+	MSD_SE_NPC_VM_CMNTW_QUESTN  = 0x88EE,
+	MSD_SE_NPC_VM_CMNTW_THANKS  = 0x88EF,
+	MSD_SE_NPC_VA_PEACH_CRY1    = 0x88F0,
+	MSD_SE_NPC_VA_PEACH_CRY2    = 0x88F1,
+	MSD_SE_NPC_VA_PEACH_CRY3    = 0x88F2,
+	MSD_SE_NPC_VA_PEACH_BREATH1 = 0x88F3,
+	MSD_SE_NPC_VA_PEACH_BREATH2 = 0x88F4,
+	MSD_SE_NPC_VA_PEACH_WET     = 0x88F5,
+	MSD_SE_NPC_VA_KINOJ_CRY1    = 0x88F6,
+	MSD_SE_NPC_VA_KINOJ_CRY2    = 0x88F7,
+	MSD_SE_NPC_VA_KINOJ_WET1    = 0x88F8,
+	MSD_SE_NPC_VA_KINOJ_WET2    = 0x88F9,
+	MSD_SE_NPC_VA_KINOP_WET1    = 0x88FA,
+	MSD_SE_NPC_VA_KINOP_WET2    = 0x88FB,
+	MSD_SE_NPC_VA_KINOP_ANDO    = 0x88FC,
+	MSD_SE_NPC_VA_KINOP_PRESS1  = 0x88FD,
+	MSD_SE_NPC_VA_KINOP_PRESS2  = 0x88FE,
+	MSD_SE_NPC_VA_KINOP_JOY     = 0x88FF,
+	MSD_SE_NPC_VA_KINOP_CONFUSE = 0x8900,
+	MSD_SE_NPC_VA_MONTM_KUSHAMI = 0x8901,
+	MSD_SE_NPC_VA_MONTM_ANGRY1  = 0x8902,
+	MSD_SE_NPC_VA_MONTM_ANGRY2  = 0x8903,
+	MSD_SE_NPC_VA_MONTM_SINK    = 0x8904,
+	MSD_SE_NPC_VA_MONTM_APEAR   = 0x8905,
+	MSD_SE_NPC_VA_MONTM_TROUB   = 0x8906,
+	MSD_SE_NPC_VA_MONTM_THROW   = 0x8907,
+	MSD_SE_NPC_VA_MONTM_PRESS1  = 0x8908,
+	MSD_SE_NPC_VA_MONTM_PRESS2  = 0x8909,
+	MSD_SE_NPC_VA_MONTM_PRESS3  = 0x890A,
+	MSD_SE_NPC_VA_MONTM_PRESS4  = 0x890B,
+	MSD_SE_NPC_VA_MONTM_DANCE1  = 0x890C,
+	MSD_SE_NPC_VA_MONTM_DANCE2  = 0x890D,
+	MSD_SE_NPC_VA_MONTM_GOOD1   = 0x890E,
+	MSD_SE_NPC_VA_MONTM_GOOD2   = 0x890F,
+	MSD_SE_NPC_VA_MONTM_RUNNING = 0x8910,
+	MSD_SE_NPC_VA_MONTW_KUSHAMI = 0x8911,
+	MSD_SE_NPC_VA_MONTW_KUSHAM2 = 0x8912,
+	MSD_SE_NPC_VA_MONTW_ANGRY1  = 0x8913,
+	MSD_SE_NPC_VA_MONTW_SINK    = 0x8914,
+	MSD_SE_NPC_VA_MONTW_ANDO1   = 0x8915,
+	MSD_SE_NPC_VA_MONTW_ANDO2   = 0x8916,
+	MSD_SE_NPC_VA_MONTW_TROUB1  = 0x8917,
+	MSD_SE_NPC_VA_MONTW_TROUB2  = 0x8918,
+	MSD_SE_NPC_VA_MONTW_THROW1  = 0x8919,
+	MSD_SE_NPC_VA_MONTW_PRESS1  = 0x891A,
+	MSD_SE_NPC_VA_MONTW_PRESS2  = 0x891B,
+	MSD_SE_NPC_VA_MONTW_ATTENT  = 0x891C,
+	MSD_SE_NPC_VA_MONTW_DANCE1  = 0x891D,
+	MSD_SE_NPC_VA_MONTW_DANCE2  = 0x891E,
+	MSD_SE_NPC_VA_MONTW_RUNNING = 0x891F,
+	MSD_SE_BS_UNG_VACUUM        = 0x8120,
+	MSD_SE_BS_UNG_ROLL          = 0x8921,
+	MSD_SE_BS_UNG_UP            = 0x8922,
+	MSD_SE_BS_UNG_DOWN          = 0x8923,
+	MSD_SE_BS_UNG_BITE          = 0x8924,
+	MSD_SE_BS_UNG_THROW_UP      = 0x8925,
+	MSD_SE_BS_UNG_TEAR_TREMBLE  = 0x8926,
+	MSD_SE_BS_UNG_TEAR_CLASH    = 0x8927,
+	MSD_SE_BS_UNG_TEATH_COMEOFF = 0x8928,
+	MSD_SE_BS_UNG_TEATH_FLASH   = 0x8929,
+	MSD_SE_BS_UNG_VOICE_M_CRY   = 0x892A,
+	MSD_SE_BS_UNG_VOICE_W_CRY   = 0x892B,
+	MSD_SE_BS_UNG_VOICE_M_JOY   = 0x892C,
+	MSD_SE_BS_UNG_VOICE_W_JOY   = 0x892D,
+	MSD_SE_BS_UNG_VOICE_NORMAL  = 0x812E,
+	MSD_SE_BS_UNG_VOICE_LAST    = 0x892F,
+	MSD_SE_BS_MKP_BREAK         = 0x8930,
+	MSD_SE_BS_MKP_MOVEMENT      = 0x8931,
+	MSD_SE_BS_MKP_BOUND         = 0x8932,
+	MSD_SE_BS_MKP_EXPLOSION_S   = 0x8933,
+	MSD_SE_BS_MKP_SPARK         = 0x8934,
+	MSD_SE_BS_MKP_FIRE          = 0x8135,
+	MSD_SE_BS_MKP_SPARK_LV_1    = 0x8136,
+	MSD_SE_BS_MKP_SPARK_LV_2    = 0x8137,
+	MSD_SE_BS_MKP_SPARK_LV_3    = 0x8138,
+	MSD_SE_BS_MKP_SPARK_LV_4    = 0x8139,
+	MSD_SE_BS_UNG_VOICE_NDUMMY0 = 0x813D,
+	MSD_SE_BS_UNG_VOICE_NDUMMY1 = 0x813E,
+	MSD_SE_BS_UNG_MOUTH_CLOSE   = 0x893F,
+	MSD_SE_NPC_VA_MAREM_JUMP1   = 0x8940,
+	MSD_SE_NPC_VA_MAREM_JUMP2   = 0x8941,
+	MSD_SE_NPC_VA_MAREM_SINK    = 0x8942,
+	MSD_SE_NPC_VA_MAREM_ANDO    = 0x8943,
+	MSD_SE_NPC_VA_MAREM_CONFUSE = 0x8944,
+	MSD_SE_NPC_VA_MAREM_PRESS   = 0x8945,
+	MSD_SE_NPC_VA_MAREM_DANCE1  = 0x8946,
+	MSD_SE_NPC_VA_MAREM_DANCE2  = 0x8947,
+	MSD_SE_NPC_VA_MAREM_HUKKI   = 0x8948,
+	MSD_SE_NPC_VA_MAREW_ANGRY   = 0x8949,
+	MSD_SE_NPC_VA_MAREW_JUMP    = 0x894A,
+	MSD_SE_NPC_VA_MAREW_SINK    = 0x894B,
+	MSD_SE_NPC_VA_MAREW_ANDO    = 0x894C,
+	MSD_SE_NPC_VA_MAREW_CONFUSE = 0x894D,
+	MSD_SE_NPC_VA_MAREW_PRESS   = 0x894E,
+	MSD_SE_NPC_VA_MAREW_DANCE1  = 0x894F,
+	MSD_SE_NPC_VA_MAREW_DANCE2  = 0x8950,
+	MSD_SE_NPC_VA_MAREW_HUKKI   = 0x8951,
+	MSD_SE_NPC_VA_MAREJ_WET     = 0x8952,
+	MSD_SE_NPC_VA_MAREJ_PRESS1  = 0x8953,
+	MSD_SE_NPC_VA_MAREJ_PRESS2  = 0x8954,
+	MSD_SE_NPC_VA_MAREJ_DANCE1  = 0x8955,
+	MSD_SE_NPC_VA_MAREJ_DANCE2  = 0x8956,
+	MSD_SE_NPC_VA_MAREB_LAUGH   = 0x8957,
+	MSD_SE_NPC_VA_MAREB_USUAL   = 0x8958,
+	MSD_SE_NPC_VA_MAREB_CRY1    = 0x8959,
+	MSD_SE_NPC_VA_MAREB_CRY2    = 0x895A,
+	MSD_SE_NPC_VA_CMNTM_KUSHAMI = 0x895B,
+	MSD_SE_NPC_VA_CMNTM_ANGRY1  = 0x895C,
+	MSD_SE_NPC_VA_CMNTM_ANGRY2  = 0x895D,
+	MSD_SE_NPC_VA_CMNTM_SINK    = 0x895E,
+	MSD_SE_NPC_VA_CMNTM_APEAR   = 0x895F,
+	MSD_SE_NPC_VA_CMNTM_TROUB   = 0x8960,
+	MSD_SE_NPC_VA_CMNTM_THROW   = 0x8961,
+	MSD_SE_NPC_VA_CMNTM_PRESS1  = 0x8962,
+	MSD_SE_NPC_VA_CMNTM_PRESS2  = 0x8963,
+	MSD_SE_NPC_VA_CMNTM_PRESS3  = 0x8964,
+	MSD_SE_NPC_VA_CMNTM_PRESS4  = 0x8965,
+	MSD_SE_NPC_VA_CMNTM_DANCE1  = 0x8966,
+	MSD_SE_NPC_VA_CMNTM_DANCE2  = 0x8967,
+	MSD_SE_NPC_VA_CMNTM_GOOD1   = 0x8968,
+	MSD_SE_NPC_VA_CMNTM_GOOD2   = 0x8969,
+	MSD_SE_NPC_VA_CMNTM_RUNNING = 0x896A,
+	MSD_SE_NPC_VA_CMNTW_KUSHAMI = 0x896B,
+	MSD_SE_NPC_VA_CMNTW_KUSHAM2 = 0x896C,
+	MSD_SE_NPC_VA_CMNTW_ANGRY1  = 0x896D,
+	MSD_SE_NPC_VA_CMNTW_SINK    = 0x896E,
+	MSD_SE_NPC_VA_CMNTW_ANDO1   = 0x896F,
+	MSD_SE_NPC_VA_CMNTW_ANDO2   = 0x8970,
+	MSD_SE_NPC_VA_CMNTW_TROUB1  = 0x8971,
+	MSD_SE_NPC_VA_CMNTW_TROUB2  = 0x8972,
+	MSD_SE_NPC_VA_CMNTW_THROW1  = 0x8973,
+	MSD_SE_NPC_VA_CMNTW_PRESS1  = 0x8974,
+	MSD_SE_NPC_VA_CMNTW_PRESS2  = 0x8975,
+	MSD_SE_NPC_VA_CMNTW_ATTENT  = 0x8976,
+	MSD_SE_NPC_VA_CMNTW_DANCE1  = 0x8977,
+	MSD_SE_NPC_VA_CMNTW_DANCE2  = 0x8978,
+	MSD_SE_NPC_VA_CMNTW_RUNNING = 0x8979,
+	MSD_SE_NPC_VA_CMAREM_JUMP1  = 0x897A,
+	MSD_SE_NPC_VA_CMAREM_JUMP2  = 0x897B,
+	MSD_SE_NPC_VA_CMAREM_SINK   = 0x897C,
+	MSD_SE_NPC_VA_CMAREM_ANDO   = 0x897D,
+	MSD_SE_NPC_VA_CMAREM_CONFUS = 0x897E,
+	MSD_SE_NPC_VA_CMAREM_PRESS  = 0x897F,
+	MSD_SE_NPC_VA_CMAREM_DANCE1 = 0x8980,
+	MSD_SE_NPC_VA_CMAREM_DANCE2 = 0x8981,
+	MSD_SE_NPC_VA_CMAREM_HUKKI  = 0x8982,
+	MSD_SE_NPC_VA_CMAREW_ANGRY  = 0x8983,
+	MSD_SE_NPC_VA_CMAREW_JUMP   = 0x8984,
+	MSD_SE_NPC_VA_CMAREW_SINK   = 0x8985,
+	MSD_SE_NPC_VA_CMAREW_ANDO   = 0x8986,
+	MSD_SE_NPC_VA_CMAREW_CONFUS = 0x8987,
+	MSD_SE_NPC_VA_CMAREW_PRESS  = 0x8988,
+	MSD_SE_NPC_VA_CMAREW_DANCE1 = 0x8989,
+	MSD_SE_NPC_VA_CMAREW_DANCE2 = 0x898A,
+	MSD_SE_NPC_VA_CMAREW_HUKKI  = 0x898B,
+	MSD_SE_NPC_VA_MTMAN_JUMP1   = 0x898C,
+	MSD_SE_NPC_VA_MTMAN_JUMP2   = 0x898D,
+	MSD_SE_BS_MANTA_DRAW_DEMO   = 0x818E,
+	MSD_SE_BS_MANTA_ALL_DEATH   = 0x898F,
+	MSD_SE_BS_MANTA_DAMAGE_1    = 0x8990,
+	MSD_SE_BS_MANTA_DAMAGE_2    = 0x8991,
+	MSD_SE_BS_MANTA_DAMAGE_3    = 0x8992,
+	MSD_SE_BS_MANTA_DAMAGE_4    = 0x8993,
+	MSD_SE_BS_MANTA_SEGMENT_1   = 0x8994,
+	MSD_SE_BS_MANTA_SEGMENT_2   = 0x8995,
+	MSD_SE_BS_MANTA_SEGMENT_3   = 0x8996,
+	MSD_SE_BS_MANTA_SEGMENT_4   = 0x8997,
+	MSD_SE_BS_MANTA_VANISH      = 0x8998,
+	MSD_SE_BS_MANTA_DRAW_LV     = 0x8199,
+	MSD_SE_BS_MANTA_DAMAGE_5    = 0x899A,
+	MSD_SE_BS_MANTA_ATTACK      = 0x899B,
+	MSD_SE_BS_KOOPA_HIP_ATTACK  = 0x899C,
+	MSD_SE_BS_KOOPA_FLOOD       = 0x819D,
+	MSD_SE_BS_KOOPA_JR_ATTACK   = 0x899E,
+	MSD_SE_BS_KOOPA_JR_LAUGH    = 0x899F,
+	MSD_SE_BS_KOOPA_WALK        = 0x89A0,
+	MSD_SE_BS_KOOPA_BOUND       = 0x89A1,
+	MSD_SE_BS_KOOPA_FIRE_START  = 0x89A2,
+	MSD_SE_BS_KOOPA_FIRE_LOOP   = 0x81A3,
+	MSD_SE_BS_KOOPA_VO_THREAT_0 = 0x89A4,
+	MSD_SE_BS_KOOPA_VO_THREAT_1 = 0x89A5,
+	MSD_SE_BS_KOOPA_VO_DAMAGE_0 = 0x89A6,
+	MSD_SE_BS_KOOPA_VO_DAMAGE_1 = 0x89A7,
+	MSD_SE_BS_KOOPA_PREPARE     = 0x89A8,
+	MSD_SE_BS_KOOPA_WASH        = 0x89A9,
+	MSD_SE_BS_KOOPA_VO_GROAN_0  = 0x89AA,
+	MSD_SE_BS_KOOPA_VO_GROAN_1  = 0x89AB,
+	MSD_SE_BS_KOOPA_VO_CRY      = 0x89AC,
+	MSD_SE_BS_KOOPA_VO_LAUGH    = 0x89AD,
+	MSD_SE_BS_KOOPA_WATER_DIVE  = 0x89AE,
+	MSD_SE_BS_FALLWATER         = 0x89AF,
+	MSD_SE_BS_KOOPA_JR_DAMAGE   = 0x89B0,
+	MSD_SE_SUPER_HIP_DROP_FALL  = 0x81B1,
+	MSD_SE_OBJ_BI_BELL          = 0x89B8,
+	MSD_SE_OBJ_FLOWER_OPEN_0    = 0x89B9,
+	MSD_SE_OBJ_FLOWER_OPEN_1    = 0x89BA,
+	MSD_SE_OBJ_FLOWER_OPEN_2    = 0x89BB,
+	MSD_SE_OBJ_FLOWER_OPEN_3    = 0x89BC,
+	MSD_SE_OBJ_FLOWER_OPEN_4    = 0x89BD,
+	MSD_SE_OBJ_FLOWER_OPEN_5    = 0x89BE,
+	MSD_SE_OBJ_FLOWER_OPEN_6    = 0x89BF,
+	MSD_SE_OBJ_FLOWER_OPEN_7    = 0x89C0,
+	MSD_SE_SHINE_EXIST          = 0x81C1,
+};
